### 2023-11-20 4d12c92
test mixing funcs; Google docstrings
