+++ src/pyEQL/functions.py
@@ -10,51 +10,44 @@ from typing import Literal
 
 from monty.dev import deprecated
 
-import pyEQL
-from pyEQL import ureg
+from pyEQL import Solution, ureg
 from pyEQL.logging_system import logger
 
 
-def gibbs_mix(Solution1, Solution2):
+def gibbs_mix(solution1: Solution, solution2: Solution):
     r"""
     Return the Gibbs energy change associated with mixing two solutions.
 
-    Parameters
-    ----------
-    Solution1, Solution2 : Solution objects
-        The two solutions to be mixed.
+    Args:
+        solution1, solution2: The two solutions to be mixed.
 
-    Returns
-    -------
-    Quantity
+    Returns:
         The change in Gibbs energy associated with complete mixing of the
         Solutions, in Joules.
 
-    Notes
-    -----
-    The Gibbs energy of mixing is calculated as follows
+    Notes:
+        The Gibbs energy of mixing is calculated as follows
 
-    .. math::
+        .. math::
 
-        \\Delta_{mix} G = \\sum_i (n_c + n_d) R T \\ln a_b - \\sum_i n_c R T \\ln a_c - \\sum_i n_d R T \\ln a_d
+            \\Delta_{mix} G = \\sum_i (n_c + n_d) R T \\ln a_b - \\sum_i n_c R T \\ln a_c - \\sum_i n_d R T \\ln a_d
 
-    Where :math:`n` is the number of moles of substance, :math:`T` is the temperature in kelvin,
-    and  subscripts :math:`b`, :math:`c`, and :math:`d` refer to the concentrated, dilute, and blended
-    Solutions, respectively.
+        Where :math:`n` is the number of moles of substance, :math:`T` is the temperature in kelvin,
+        and  subscripts :math:`b`, :math:`c`, and :math:`d` refer to the concentrated, dilute, and blended
+        Solutions, respectively.
 
-    Note that dissociated ions must be counted as separate components,
-    so a simple salt dissolved in water is a three component solution (cation,
-    anion, and water).
+        Note that dissociated ions must be counted as separate components,
+        so a simple salt dissolved in water is a three component solution (cation,
+        anion, and water).
 
-    References
-    ----------
-    Koga, Yoshikata, 2007. *Solution Thermodynamics and its Application to Aqueous Solutions:
-        A differential approach.* Elsevier, 2007, pp. 23-37.
+    References:
+        Koga, Yoshikata, 2007. *Solution Thermodynamics and its Application to Aqueous Solutions:
+            A differential approach.* Elsevier, 2007, pp. 23-37.
 
     """
-    concentrate = Solution1
-    dilute = Solution2
-    blend = Solution1 + Solution2
+    concentrate = solution1
+    dilute = solution2
+    blend = solution1 + solution2
     term_list = {concentrate: 0, dilute: 0, blend: 0}
 
     # calculate the entropy change and number of moles solute for each solution
@@ -68,46 +61,40 @@ def gibbs_mix(Solution1, Solution2):
     )
 
 
-def entropy_mix(Solution1, Solution2):
+def entropy_mix(solution1: Solution, solution2: Solution):
     r"""
     Return the ideal mixing entropy associated with mixing two solutions.
 
-    Parameters
-    ----------
-    Solution1, Solution2 : Solution objects
-        The two solutions to be mixed.
+    Parameters:
+        solution1, solution2: The two solutions to be mixed.
 
-    Returns
-    -------
-    Quantity
+    Returns:
         The ideal mixing entropy associated with complete mixing of the
         Solutions, in Joules.
 
-    Notes
-    -----
-    The ideal entropy of mixing is calculated as follows
+    Notes:
+        The ideal entropy of mixing is calculated as follows
 
-    .. math::
+        .. math::
 
[…]
