### 2023-11-20 81cfa88
autogenerate deprecation fixes
