### 2023-11-21 4cf17f4
undo adding requests in setup.cfg; remove minVersion key from tox.ini
