+++ setup.cfg
@@ -78,7 +78,6 @@ testing =
     mypy
     ruff
     tox
-    requests
 docs =
     sphinx>=3.2.1
     sphinx-rtd-theme
