### 2023-11-21 dc4dbb1
set tox minVersion to 4.11; update mypy config in toml
