+++ tox.ini
@@ -3,6 +3,7 @@
 # THIS SCRIPT IS SUPPOSED TO BE AN EXAMPLE. MODIFY IT ACCORDING TO YOUR NEEDS!
 
 [tox]
+minversion = 4.11
 envlist = {py39,py310,py11,py312}-{linux,windows,macos}
 
 
