### 2023-11-22 1d8940b
set minversion & update tox dependency
