+++ setup.cfg
@@ -77,7 +77,7 @@ testing =
     black
     mypy
     ruff
-    tox
+    tox<4
 docs =
     sphinx>=3.2.1
     sphinx-rtd-theme
