### 2023-11-24 df55c37
updated min pytest version; updated ruff config naming; added config to ignore warnings for pytest
