+++ pyproject.toml
@@ -10,7 +10,7 @@ version_scheme = "no-guess-dev"
 [tool.black]
 line-length = 120
 
-[tool.ruff]
+[tool.ruff.lint]
 target-version = "py38"
 line-length = 120
 exclude = [".dat"]
@@ -68,12 +68,12 @@ ignore = [
 pydocstyle.convention = "google"
 isort.split-on-trailing-comma = false
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]
 "tests/*" = ["D", "E501"] # TODO - remove E501
 
 [tool.pytest.ini_options]
-minversion = "3.0"
+minversion = "6.0"
 addopts = [
   "-ra",
   "--showlocals",
@@ -88,6 +88,9 @@ filterwarnings = ["error"]
 log_cli_level = "warning"
 testpaths = [
   "tests",
+]
+filterwarnings = [
+  "ignore::UserWarning",
 ]
 
 [tool.mypy]
