+++ pyproject.toml
@@ -89,9 +89,7 @@ log_cli_level = "warning"
 testpaths = [
   "tests",
 ]
-filterwarnings = [
-  "ignore::UserWarning",
-]
+filterwarnings = ["ignore::UserWarning"]
 
 [tool.mypy]
 ignore_missing_imports = true
