### 2023-11-24 e12a58e
updating setup.cfg to ignore warnings
