+++ setup.cfg
@@ -110,6 +110,8 @@ norecursedirs =
     build
     .tox
 testpaths = tests
+filterwarnings =
+    ignore::UserWarning
 # Use pytest markers to select/deselect specific tests
 # markers =
 #     slow: mark tests as slow (deselect with '-m "not slow"')
