### 2023-11-24 f22b900
remove filterwarning from pyproject
