### 2023-11-24 3959d01
add filterwarnings but correctly this time
