+++ requirements/macos-latest_py3.9_extras.txt
@@ -60,7 +60,7 @@ coverage[toml]==7.3.2
     # via
     #   coverage
     #   pytest-cov
-cryptography==41.0.5
+cryptography==41.0.6
     # via paramiko
 cycler==0.12.1
     # via matplotlib
