+++ requirements/ubuntu-latest_py3.11_extras.txt
@@ -20,21 +20,19 @@ attrs==23.1.0
     #   referencing
 babel==2.13.1
     # via sphinx
-bcrypt==4.0.1
+bcrypt==4.1.1
     # via paramiko
 black==23.11.0
     # via pyEQL (setup.py)
 blinker==1.7.0
     # via flask
-boto3==1.28.83
+boto3==1.33.11
     # via maggma
-botocore==1.31.83
+botocore==1.33.11
     # via
     #   boto3
     #   s3transfer
-cachetools==5.3.2
-    # via tox
-certifi==2023.7.22
+certifi==2023.11.17
     # via requests
 cffi==1.16.0
     # via
@@ -42,8 +40,6 @@ cffi==1.16.0
     #   pynacl
 cfgv==3.4.0
     # via pre-commit
-chardet==5.2.0
-    # via tox
 charset-normalizer==3.3.2
     # via requests
 click==8.1.7
@@ -52,15 +48,13 @@ click==8.1.7
     #   flask
     #   mongogrant
     #   uvicorn
-colorama==0.4.6
-    # via tox
 contourpy==1.2.0
     # via matplotlib
 coverage[toml]==7.3.2
     # via
     #   coverage
     #   pytest-cov
-cryptography==41.0.5
+cryptography==41.0.7
     # via paramiko
 cycler==0.12.1
     # via matplotlib
@@ -70,7 +64,7 @@ dnspython==2.4.2
     # via
     #   maggma
     #   pymongo
-docutils==0.18.1
+docutils==0.20.1
     # via
     #   myst-parser
     #   sphinx
@@ -83,7 +77,7 @@ filelock==3.13.1
     #   virtualenv
 flask==3.0.0
     # via mongogrant
-fonttools==4.44.0
+fonttools==4.46.0
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -91,9 +85,9 @@ h11==0.14.0
     # via uvicorn
 iapws==1.5.3
     # via pyEQL (setup.py)
-identify==2.5.31
+identify==2.5.33
     # via pre-commit
-idna==3.4
+idna==3.6
     # via
     #   anyio
     #   requests
@@ -114,9 +108,9 @@ jmespath==1.0.1
     #   botocore
 joblib==1.3.2
     # via pymatgen
-jsonschema==4.19.2
+jsonschema==4.20.0
     # via maggma
-jsonschema-specifications==2023.7.1
+jsonschema-specifications==2023.11.2
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
@@ -124,7 +118,7 @@ latexcodec==2.0.1
     # via pybtex
 linkify-it-py==2.0.2
     # via myst-parser
-maggma==0.57.8
+maggma==0.58.0
     # via pyEQL (setup.py)
 markdown-it-py==3.0.0
     # via
@@ -135,7 +129,7 @@ markupsafe==2.1.3
     # via
     #   jinja2
     #   werkzeug
-matplotlib==3.8.1
+matplotlib==3.8.2
     # via pymatgen
 mdit-py-plugins==0.4.0
     # via myst-parser
@@ -154,7 +148,7 @@ mpmath==1.3.0
     # via sympy
 msgpack==1.0.7
     # via maggma
-mypy==1.6.1
+mypy==1.7.1
     # via pyEQL (setup.py)
 mypy-extensions==1.0.0
     # via
@@ -166,7 +160,7 @@ networkx==3.2.1
     # via pymatgen
 nodeenv==1.8.0
     # via pre-commit
-numpy==1.26.1
+numpy==1.26.2
     # via
     #   contourpy
     #   maggma
[…]
