+++ requirements/macos-latest_py3.12.txt
@@ -1,0 +1,256 @@
+#
+# This file is autogenerated by pip-compile with Python 3.12
+# by the following command:
+#
+#    pip-compile --output-file=requirements/macos-latest_py3.12.txt
+#
+aioitertools==0.11.0
+    # via maggma
+annotated-types==0.6.0
+    # via pydantic
+anyio==3.7.1
+    # via
+    #   fastapi
+    #   starlette
+attrs==23.1.0
+    # via
+    #   jsonschema
+    #   referencing
+bcrypt==4.1.1
+    # via paramiko
+blinker==1.7.0
+    # via flask
+boto3==1.33.11
+    # via maggma
+botocore==1.33.11
+    # via
+    #   boto3
+    #   s3transfer
+certifi==2023.11.17
+    # via requests
+cffi==1.16.0
+    # via
+    #   cryptography
+    #   pynacl
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via
+    #   flask
+    #   mongogrant
+    #   uvicorn
+contourpy==1.2.0
+    # via matplotlib
+cryptography==41.0.7
+    # via paramiko
+cycler==0.12.1
+    # via matplotlib
+dnspython==2.4.2
+    # via
+    #   maggma
+    #   pymongo
+fastapi==0.104.1
+    # via maggma
+flask==3.0.0
+    # via mongogrant
+fonttools==4.46.0
+    # via matplotlib
+future==0.18.3
+    # via uncertainties
+h11==0.14.0
+    # via uvicorn
+iapws==1.5.3
+    # via pyEQL (setup.py)
+idna==3.6
+    # via
+    #   anyio
+    #   requests
+itsdangerous==2.1.2
+    # via flask
+jinja2==3.1.2
+    # via flask
+jmespath==1.0.1
+    # via
+    #   boto3
+    #   botocore
+joblib==1.3.2
+    # via pymatgen
+jsonschema==4.20.0
+    # via maggma
+jsonschema-specifications==2023.11.2
+    # via jsonschema
+kiwisolver==1.4.5
+    # via matplotlib
+latexcodec==2.0.1
+    # via pybtex
+maggma==0.58.0
+    # via pyEQL (setup.py)
+markupsafe==2.1.3
+    # via
+    #   jinja2
+    #   werkzeug
+matplotlib==3.8.2
+    # via pymatgen
+mongogrant==0.3.3
+    # via maggma
+mongomock==4.1.2
+    # via maggma
+monty==2023.11.3
+    # via
+    #   maggma
+    #   pyEQL (setup.py)
+    #   pymatgen
+mpmath==1.3.0
+    # via sympy
+msgpack==1.0.7
+    # via maggma
+networkx==3.2.1
+    # via pymatgen
+numpy==1.26.2
+    # via
+    #   contourpy
+    #   maggma
+    #   matplotlib
+    #   pandas
+    #   periodictable
+    #   pyEQL (setup.py)
+    #   pymatgen
+    #   scipy
+    #   spglib
+orjson==3.9.10
+    # via maggma
+packaging==23.2
+    # via
+    #   matplotlib
+    #   mongomock
+    #   plotly
+palettable==3.3.3
+    # via pymatgen
+pandas==2.1.4
+    # via pymatgen
+paramiko==3.3.1
+    # via sshtunnel
+periodictable==1.6.1
+    # via phreeqpython
+phreeqpython==1.5.0
+    # via pyEQL (setup.py)
+pillow==10.1.0
+    # via matplotlib
+pint==0.23
+    # via pyEQL (setup.py)
+plotly==5.18.0
+    # via pymatgen
+pybtex==0.24.0
+    # via pymatgen
[…]
