### 2023-12-10 ad52250
update dependencies for . (ubuntu-latest/py3.10)
