+++ requirements/ubuntu-latest_py3.10_extras.txt
@@ -20,21 +20,19 @@ attrs==23.1.0
     #   referencing
 babel==2.13.1
     # via sphinx
-bcrypt==4.0.1
+bcrypt==4.1.1
     # via paramiko
 black==23.11.0
     # via pyEQL (setup.py)
 blinker==1.7.0
     # via flask
-boto3==1.28.83
+boto3==1.33.11
     # via maggma
-botocore==1.31.83
+botocore==1.33.11
     # via
     #   boto3
     #   s3transfer
-cachetools==5.3.2
-    # via tox
-certifi==2023.7.22
+certifi==2023.11.17
     # via requests
 cffi==1.16.0
     # via
@@ -42,8 +40,6 @@ cffi==1.16.0
     #   pynacl
 cfgv==3.4.0
     # via pre-commit
-chardet==5.2.0
-    # via tox
 charset-normalizer==3.3.2
     # via requests
 click==8.1.7
@@ -52,15 +48,13 @@ click==8.1.7
     #   flask
     #   mongogrant
     #   uvicorn
-colorama==0.4.6
-    # via tox
 contourpy==1.2.0
     # via matplotlib
 coverage[toml]==7.3.2
     # via
     #   coverage
     #   pytest-cov
-cryptography==41.0.5
+cryptography==41.0.7
     # via paramiko
 cycler==0.12.1
     # via matplotlib
@@ -70,12 +64,12 @@ dnspython==2.4.2
     # via
     #   maggma
     #   pymongo
-docutils==0.18.1
+docutils==0.20.1
     # via
     #   myst-parser
     #   sphinx
     #   sphinx-rtd-theme
-exceptiongroup==1.1.3
+exceptiongroup==1.2.0
     # via
     #   anyio
     #   pytest
@@ -87,7 +81,7 @@ filelock==3.13.1
     #   virtualenv
 flask==3.0.0
     # via mongogrant
-fonttools==4.44.0
+fonttools==4.46.0
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -95,9 +89,9 @@ h11==0.14.0
     # via uvicorn
 iapws==1.5.3
     # via pyEQL (setup.py)
-identify==2.5.31
+identify==2.5.33
     # via pre-commit
-idna==3.4
+idna==3.6
     # via
     #   anyio
     #   requests
@@ -118,9 +112,9 @@ jmespath==1.0.1
     #   botocore
 joblib==1.3.2
     # via pymatgen
-jsonschema==4.19.2
+jsonschema==4.20.0
     # via maggma
-jsonschema-specifications==2023.7.1
+jsonschema-specifications==2023.11.2
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
@@ -128,7 +122,7 @@ latexcodec==2.0.1
     # via pybtex
 linkify-it-py==2.0.2
     # via myst-parser
-maggma==0.57.8
+maggma==0.58.0
     # via pyEQL (setup.py)
 markdown-it-py==3.0.0
     # via
@@ -139,7 +133,7 @@ markupsafe==2.1.3
     # via
     #   jinja2
     #   werkzeug
-matplotlib==3.8.1
+matplotlib==3.8.2
     # via pymatgen
 mdit-py-plugins==0.4.0
     # via myst-parser
@@ -158,7 +152,7 @@ mpmath==1.3.0
     # via sympy
 msgpack==1.0.7
     # via maggma
-mypy==1.6.1
+mypy==1.7.1
     # via pyEQL (setup.py)
 mypy-extensions==1.0.0
     # via
@@ -170,7 +164,7 @@ networkx==3.2.1
     # via pymatgen
 nodeenv==1.8.0
[…]
