### 2023-12-10 fea47b9
update dependencies for . (ubuntu-latest/py3.12)
