+++ requirements/macos-latest_py3.11.txt
@@ -16,17 +16,17 @@ attrs==23.1.0
     # via
     #   jsonschema
     #   referencing
-bcrypt==4.0.1
+bcrypt==4.1.1
     # via paramiko
 blinker==1.7.0
     # via flask
-boto3==1.28.83
+boto3==1.33.11
     # via maggma
-botocore==1.31.83
+botocore==1.33.11
     # via
     #   boto3
     #   s3transfer
-certifi==2023.7.22
+certifi==2023.11.17
     # via requests
 cffi==1.16.0
     # via
@@ -41,7 +41,7 @@ click==8.1.7
     #   uvicorn
 contourpy==1.2.0
     # via matplotlib
-cryptography==41.0.5
+cryptography==41.0.7
     # via paramiko
 cycler==0.12.1
     # via matplotlib
@@ -53,7 +53,7 @@ fastapi==0.104.1
     # via maggma
 flask==3.0.0
     # via mongogrant
-fonttools==4.44.0
+fonttools==4.46.0
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -61,7 +61,7 @@ h11==0.14.0
     # via uvicorn
 iapws==1.5.3
     # via pyEQL (setup.py)
-idna==3.4
+idna==3.6
     # via
     #   anyio
     #   requests
@@ -75,21 +75,21 @@ jmespath==1.0.1
     #   botocore
 joblib==1.3.2
     # via pymatgen
-jsonschema==4.19.2
+jsonschema==4.20.0
     # via maggma
-jsonschema-specifications==2023.7.1
+jsonschema-specifications==2023.11.2
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
 latexcodec==2.0.1
     # via pybtex
-maggma==0.57.8
+maggma==0.58.0
     # via pyEQL (setup.py)
 markupsafe==2.1.3
     # via
     #   jinja2
     #   werkzeug
-matplotlib==3.8.1
+matplotlib==3.8.2
     # via pymatgen
 mongogrant==0.3.3
     # via maggma
@@ -106,7 +106,7 @@ msgpack==1.0.7
     # via maggma
 networkx==3.2.1
     # via pymatgen
-numpy==1.26.1
+numpy==1.26.2
     # via
     #   contourpy
     #   maggma
@@ -126,7 +126,7 @@ packaging==23.2
     #   plotly
 palettable==3.3.3
     # via pymatgen
-pandas==2.1.2
+pandas==2.1.4
     # via pymatgen
 paramiko==3.3.1
     # via sshtunnel
@@ -136,7 +136,7 @@ phreeqpython==1.5.0
     # via pyEQL (setup.py)
 pillow==10.1.0
     # via matplotlib
-pint==0.22
+pint==0.23
     # via pyEQL (setup.py)
 plotly==5.18.0
     # via pymatgen
@@ -144,20 +144,20 @@ pybtex==0.24.0
     # via pymatgen
 pycparser==2.21
     # via cffi
-pydantic==2.4.2
+pydantic==2.5.2
     # via
     #   fastapi
     #   maggma
     #   pydantic-settings
-pydantic-core==2.10.1
+pydantic-core==2.14.5
     # via pydantic
-pydantic-settings==2.0.3
+pydantic-settings==2.1.0
     # via maggma
 pydash==7.0.6
     # via maggma
-pymatgen==2023.10.11
+pymatgen==2023.11.12
     # via pyEQL (setup.py)
-pymongo==4.6.0
+pymongo==4.6.1
     # via
     #   maggma
     #   mongogrant
@@ -179,9 +179,9 @@ pytz==2023.3.post1
[…]
