### 2023-12-10 bb84fd6
update dependencies for . (windows-latest/py3.12)
