+++ requirements/windows-latest_py3.12_extras.txt
@@ -1,0 +1,392 @@
+#
+# This file is autogenerated by pip-compile with Python 3.12
+# by the following command:
+#
+#    pip-compile --all-extras --output-file=requirements/windows-latest_py3.12_extras.txt
+#
+aioitertools==0.11.0
+    # via maggma
+alabaster==0.7.13
+    # via sphinx
+annotated-types==0.6.0
+    # via pydantic
+anyio==3.7.1
+    # via
+    #   fastapi
+    #   starlette
+attrs==23.1.0
+    # via
+    #   jsonschema
+    #   referencing
+babel==2.13.1
+    # via sphinx
+bcrypt==4.1.1
+    # via paramiko
+black==23.11.0
+    # via pyEQL (setup.py)
+blinker==1.7.0
+    # via flask
+boto3==1.33.11
+    # via maggma
+botocore==1.33.11
+    # via
+    #   boto3
+    #   s3transfer
+certifi==2023.11.17
+    # via requests
+cffi==1.16.0
+    # via
+    #   cryptography
+    #   pynacl
+cfgv==3.4.0
+    # via pre-commit
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via
+    #   black
+    #   flask
+    #   mongogrant
+    #   uvicorn
+colorama==0.4.6
+    # via
+    #   click
+    #   pytest
+    #   sphinx
+    #   tox
+    #   tqdm
+contourpy==1.2.0
+    # via matplotlib
+coverage[toml]==7.3.2
+    # via
+    #   coverage
+    #   pytest-cov
+cryptography==41.0.7
+    # via paramiko
+cycler==0.12.1
+    # via matplotlib
+distlib==0.3.7
+    # via virtualenv
+dnspython==2.4.2
+    # via
+    #   maggma
+    #   pymongo
+docutils==0.20.1
+    # via
+    #   myst-parser
+    #   sphinx
+    #   sphinx-rtd-theme
+fastapi==0.104.1
+    # via maggma
+filelock==3.13.1
+    # via
+    #   tox
+    #   virtualenv
+flask==3.0.0
+    # via mongogrant
+fonttools==4.46.0
+    # via matplotlib
+future==0.18.3
+    # via uncertainties
+h11==0.14.0
+    # via uvicorn
+iapws==1.5.3
+    # via pyEQL (setup.py)
+identify==2.5.33
+    # via pre-commit
+idna==3.6
+    # via
+    #   anyio
+    #   requests
+imagesize==1.4.1
+    # via sphinx
+iniconfig==2.0.0
+    # via pytest
+itsdangerous==2.1.2
+    # via flask
+jinja2==3.1.2
+    # via
+    #   flask
+    #   myst-parser
+    #   sphinx
+jmespath==1.0.1
+    # via
+    #   boto3
+    #   botocore
+joblib==1.3.2
+    # via pymatgen
+jsonschema==4.20.0
+    # via maggma
+jsonschema-specifications==2023.11.2
+    # via jsonschema
+kiwisolver==1.4.5
+    # via matplotlib
+latexcodec==2.0.1
+    # via pybtex
+linkify-it-py==2.0.2
+    # via myst-parser
+maggma==0.58.0
+    # via pyEQL (setup.py)
+markdown-it-py==3.0.0
+    # via
+    #   mdit-py-plugins
+    #   myst-parser
+    #   rich
+markupsafe==2.1.3
+    # via
+    #   jinja2
+    #   werkzeug
+matplotlib==3.8.2
+    # via pymatgen
+mdit-py-plugins==0.4.0
[…]
