+++ requirements/windows-latest_py3.11_extras.txt
@@ -193,7 +193,7 @@ palettable==3.3.3
     # via pymatgen
 pandas==2.1.4
     # via pymatgen
-paramiko==3.3.1
+paramiko==3.4.0
     # via sshtunnel
 pathspec==0.12.0
     # via black
