### 2024-01-05 b9e3207
pkg_resources -> importlib.resources
