### 2024-01-05 b3e32f1
ignore DeprecationWarning in tests
