### 2024-01-05 f7ca9b1
rm pkg_resources in from_preset
