### 2024-01-08 #6071 @lucydot
@orionarcher, @arose, @rangsimanketkaew - are you available to review this submission to JOSS?
