+++ requirements/ubuntu-latest_py3.11_extras.txt
@@ -6,29 +6,27 @@
 #
 aioitertools==0.11.0
     # via maggma
-alabaster==0.7.13
+alabaster==0.7.16
     # via sphinx
 annotated-types==0.6.0
     # via pydantic
-anyio==3.7.1
-    # via
-    #   fastapi
-    #   starlette
-attrs==23.1.0
+anyio==4.2.0
+    # via starlette
+attrs==23.2.0
     # via
     #   jsonschema
     #   referencing
-babel==2.13.1
+babel==2.14.0
     # via sphinx
-bcrypt==4.1.1
+bcrypt==4.1.2
     # via paramiko
-black==23.11.0
+black==23.12.1
     # via pyEQL (setup.py)
 blinker==1.7.0
     # via flask
-boto3==1.33.11
+boto3==1.34.15
     # via maggma
-botocore==1.33.11
+botocore==1.34.15
     # via
     #   boto3
     #   s3transfer
@@ -50,7 +48,7 @@ click==8.1.7
     #   uvicorn
 contourpy==1.2.0
     # via matplotlib
-coverage[toml]==7.3.2
+coverage[toml]==7.4.0
     # via
     #   coverage
     #   pytest-cov
@@ -58,7 +56,7 @@ cryptography==41.0.7
     # via paramiko
 cycler==0.12.1
     # via matplotlib
-distlib==0.3.7
+distlib==0.3.8
     # via virtualenv
 dnspython==2.4.2
     # via
@@ -69,7 +67,7 @@ docutils==0.20.1
     #   myst-parser
     #   sphinx
     #   sphinx-rtd-theme
-fastapi==0.104.1
+fastapi==0.108.0
     # via maggma
 filelock==3.13.1
     # via
@@ -77,7 +75,7 @@ filelock==3.13.1
     #   virtualenv
 flask==3.0.0
     # via mongogrant
-fonttools==4.46.0
+fonttools==4.47.0
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -110,7 +108,7 @@ joblib==1.3.2
     # via pymatgen
 jsonschema==4.20.0
     # via maggma
-jsonschema-specifications==2023.11.2
+jsonschema-specifications==2023.12.1
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
@@ -118,7 +116,7 @@ latexcodec==2.0.1
     # via pybtex
 linkify-it-py==2.0.2
     # via myst-parser
-maggma==0.58.0
+maggma==0.60.2
     # via pyEQL (setup.py)
 markdown-it-py==3.0.0
     # via
@@ -148,7 +146,7 @@ mpmath==1.3.0
     # via sympy
 msgpack==1.0.7
     # via maggma
-mypy==1.7.1
+mypy==1.8.0
     # via pyEQL (setup.py)
 mypy-extensions==1.0.0
     # via
@@ -160,7 +158,7 @@ networkx==3.2.1
     # via pymatgen
 nodeenv==1.8.0
     # via pre-commit
-numpy==1.26.2
+numpy==1.26.3
     # via
     #   contourpy
     #   maggma
@@ -186,15 +184,15 @@ palettable==3.3.3
     # via pymatgen
 pandas==2.1.4
     # via pymatgen
-paramiko==3.3.1
+paramiko==3.4.0
     # via sshtunnel
-pathspec==0.12.0
+pathspec==0.12.1
     # via black
 periodictable==1.6.1
     # via phreeqpython
 phreeqpython==1.5.0
     # via pyEQL (setup.py)
-pillow==10.1.0
+pillow==10.2.0
     # via matplotlib
 pint==0.23
[…]
