+++ requirements/windows-latest_py3.9.txt
@@ -8,21 +8,19 @@ aioitertools==0.11.0
     # via maggma
 annotated-types==0.6.0
     # via pydantic
-anyio==3.7.1
-    # via
-    #   fastapi
-    #   starlette
-attrs==23.1.0
+anyio==4.2.0
+    # via starlette
+attrs==23.2.0
     # via
     #   jsonschema
     #   referencing
-bcrypt==4.1.1
+bcrypt==4.1.2
     # via paramiko
 blinker==1.7.0
     # via flask
-boto3==1.33.11
+boto3==1.34.15
     # via maggma
-botocore==1.33.11
+botocore==1.34.15
     # via
     #   boto3
     #   s3transfer
@@ -55,11 +53,11 @@ dnspython==2.4.2
     #   pymongo
 exceptiongroup==1.2.0
     # via anyio
-fastapi==0.104.1
+fastapi==0.108.0
     # via maggma
 flask==3.0.0
     # via mongogrant
-fonttools==4.46.0
+fonttools==4.47.0
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -71,7 +69,7 @@ idna==3.6
     # via
     #   anyio
     #   requests
-importlib-metadata==7.0.0
+importlib-metadata==7.0.1
     # via flask
 importlib-resources==6.1.1
     # via matplotlib
@@ -87,13 +85,13 @@ joblib==1.3.2
     # via pymatgen
 jsonschema==4.20.0
     # via maggma
-jsonschema-specifications==2023.11.2
+jsonschema-specifications==2023.12.1
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
 latexcodec==2.0.1
     # via pybtex
-maggma==0.58.0
+maggma==0.60.2
     # via pyEQL (setup.py)
 markupsafe==2.1.3
     # via
@@ -116,7 +114,7 @@ msgpack==1.0.7
     # via maggma
 networkx==3.2.1
     # via pymatgen
-numpy==1.26.2
+numpy==1.26.3
     # via
     #   contourpy
     #   maggma
@@ -138,13 +136,13 @@ palettable==3.3.3
     # via pymatgen
 pandas==2.1.4
     # via pymatgen
-paramiko==3.3.1
+paramiko==3.4.0
     # via sshtunnel
 periodictable==1.6.1
     # via phreeqpython
 phreeqpython==1.5.0
     # via pyEQL (setup.py)
-pillow==10.1.0
+pillow==10.2.0
     # via matplotlib
 pint==0.23
     # via pyEQL (setup.py)
@@ -154,18 +152,18 @@ pybtex==0.24.0
     # via pymatgen
 pycparser==2.21
     # via cffi
-pydantic==2.5.2
+pydantic==2.5.3
     # via
     #   fastapi
     #   maggma
     #   pydantic-settings
-pydantic-core==2.14.5
+pydantic-core==2.14.6
     # via pydantic
 pydantic-settings==2.1.0
     # via maggma
 pydash==7.0.6
     # via maggma
-pymatgen==2023.11.12
+pymatgen==2023.12.18
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
@@ -191,7 +189,7 @@ pyyaml==6.0.1
     # via pybtex
 pyzmq==25.1.2
     # via maggma
-referencing==0.32.0
+referencing==0.32.1
     # via
     #   jsonschema
     #   jsonschema-specifications
@@ -199,7 +197,7 @@ requests==2.31.0
     # via
     #   mongogrant
     #   pymatgen
-rpds-py==0.13.2
+rpds-py==0.16.2
     # via
[…]
