+++ requirements/macos-latest_py3.10_extras.txt
@@ -101,7 +101,7 @@ iniconfig==2.0.0
     # via pytest
 itsdangerous==2.1.2
     # via flask
-jinja2==3.1.2
+jinja2==3.1.3
     # via
     #   flask
     #   myst-parser
