+++ requirements/windows-latest_py3.11.txt
@@ -71,7 +71,7 @@ idna==3.6
     #   requests
 itsdangerous==2.1.2
     # via flask
-jinja2==3.1.2
+jinja2==3.1.3
     # via flask
 jmespath==1.0.1
     # via
