+++ requirements/ubuntu-latest_py3.9.txt
@@ -73,7 +73,7 @@ importlib-resources==6.1.1
     # via matplotlib
 itsdangerous==2.1.2
     # via flask
-jinja2==3.1.2
+jinja2==3.1.3
     # via flask
 jmespath==1.0.1
     # via
