### 2024-01-13 f708688
Solution.__add__: fix static reference to class; linting
