### 2024-01-18 #6071 @orionarcher
Hi @lucydot. I should disclose that I know @rkingsbury and have provided (very minor) feedback on pyEQL in the past. If that's not an issue, I am more than happy to review this submission.
