### 2024-01-22 #6071 @lucydot
@mdavezac @JacksonBurns @yuxuanzhuang - are you available to review this submission to JOSS?
