### 2024-01-29 #6071 @lucydot
Hi @JacksonBurns @yuxuanzhuang  - excellent news :) ✨

I will add you both as reviewers and then ask editorialbot to start the review.

If you haven't reviewed for JOSS before, you can find some detailed information about the review process on the documentation pages: [reviewing for JOSS](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html). You don't need to read this all (the essential instructions are given at the top of the review thread that editorialbot will generate), but it gives some context if wanted.

The key things that are different from standard journals are: i) everything happens in a Github review thread; ii) the review is a conversation back and forth - you do not need to do you review in a single step; iii) the review is structured in that you work your way through a checklist.

We ask that reviews are completed in about 4 weeks. We advise you start the review early, as it is an iterative process between reviewers and authors.

Any questions please just ask - and thanks again for agreeing to review.
