### 2024-01-29 #6295 @editorialbot
[REVIEW]: pyEQL: A Python interface for water chemistry
**Submitting author:** @rkingsbury (<a href="http://orcid.org/0000-0002-7168-3967">Ryan Kingsbury</a>)
**Repository:** https://github.com/rkingsbury/pyEQL
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
