+++ .github/workflows/testing.yaml
@@ -59,7 +59,7 @@ jobs:
       - name: Run tests
         run: |
           pytest --cov=src/pyEQL --cov-report=xml
-      - uses: codecov/codecov-action@v3
+      - uses: codecov/codecov-action@v4
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
           file: ./coverage.xml
