### 2024-02-05 b228c06
Bump fastapi from 0.108.0 to 0.109.1 in /requirements
Bumps [fastapi](https://github.com/tiangolo/fastapi) from 0.108.0 to 0.109.1.
- [Release notes](https://github.com/tiangolo/fastapi/releases)
- [Commits](https://github.com/tiangolo/fastapi/compare/0.108.0...0.109.1)

---
updated-dependencies:
- dependency-name: fastapi
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
