+++ requirements/windows-latest_py3.12_extras.txt
@@ -74,7 +74,7 @@ docutils==0.20.1
     #   myst-parser
     #   sphinx
     #   sphinx-rtd-theme
-fastapi==0.108.0
+fastapi==0.109.1
     # via maggma
 filelock==3.13.1
     # via
