+++ requirements/macos-latest_py3.10.txt
@@ -49,7 +49,7 @@ dnspython==2.4.2
     #   pymongo
 exceptiongroup==1.2.0
     # via anyio
-fastapi==0.108.0
+fastapi==0.109.1
     # via maggma
 flask==3.0.0
     # via mongogrant
