### 2024-02-05 2c84e74
Bump starlette from 0.32.0.post1 to 0.36.2 in /requirements
Bumps [starlette](https://github.com/encode/starlette) from 0.32.0.post1 to 0.36.2.
- [Release notes](https://github.com/encode/starlette/releases)
- [Changelog](https://github.com/encode/starlette/blob/master/docs/release-notes.md)
- [Commits](https://github.com/encode/starlette/compare/0.32.0.post1...0.36.2)

---
updated-dependencies:
- dependency-name: starlette
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
