+++ requirements/windows-latest_py3.12_extras.txt
@@ -59,7 +59,7 @@ coverage[toml]==7.4.0
     # via
     #   coverage
     #   pytest-cov
-cryptography==41.0.7
+cryptography==42.0.0
     # via paramiko
 cycler==0.12.1
     # via matplotlib
