+++ requirements/macos-latest_py3.11.txt
@@ -39,7 +39,7 @@ click==8.1.7
     #   uvicorn
 contourpy==1.2.0
     # via matplotlib
-cryptography==41.0.7
+cryptography==42.0.0
     # via paramiko
 cycler==0.12.1
     # via matplotlib
