+++ requirements/ubuntu-latest_py3.9_extras.txt
@@ -20,17 +20,17 @@ babel==2.14.0
     # via sphinx
 bcrypt==4.1.2
     # via paramiko
-black==23.12.1
+black==24.1.1
     # via pyEQL (setup.py)
 blinker==1.7.0
     # via flask
-boto3==1.34.15
+boto3==1.34.35
     # via maggma
-botocore==1.34.15
+botocore==1.34.35
     # via
     #   boto3
     #   s3transfer
-certifi==2023.11.17
+certifi==2024.2.2
     # via requests
 cffi==1.16.0
     # via
@@ -48,17 +48,17 @@ click==8.1.7
     #   uvicorn
 contourpy==1.2.0
     # via matplotlib
-coverage[toml]==7.4.0
+coverage[toml]==7.4.1
     # via
     #   coverage
     #   pytest-cov
-cryptography==42.0.0
+cryptography==42.0.2
     # via paramiko
 cycler==0.12.1
     # via matplotlib
 distlib==0.3.8
     # via virtualenv
-dnspython==2.4.2
+dnspython==2.5.0
     # via
     #   maggma
     #   pymongo
@@ -71,15 +71,15 @@ exceptiongroup==1.2.0
     # via
     #   anyio
     #   pytest
-fastapi==0.109.1
+fastapi==0.109.2
     # via maggma
 filelock==3.13.1
     # via
     #   tox
     #   virtualenv
-flask==3.0.0
+flask==3.0.2
     # via mongogrant
-fonttools==4.47.0
+fonttools==4.48.1
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -105,7 +105,7 @@ iniconfig==2.0.0
     # via pytest
 itsdangerous==2.1.2
     # via flask
-jinja2==3.1.2
+jinja2==3.1.3
     # via
     #   flask
     #   myst-parser
@@ -116,7 +116,7 @@ jmespath==1.0.1
     #   botocore
 joblib==1.3.2
     # via pymatgen
-jsonschema==4.20.0
+jsonschema==4.21.1
     # via maggma
 jsonschema-specifications==2023.12.1
     # via jsonschema
@@ -124,16 +124,16 @@ kiwisolver==1.4.5
     # via matplotlib
 latexcodec==2.0.1
     # via pybtex
-linkify-it-py==2.0.2
+linkify-it-py==2.0.3
     # via myst-parser
-maggma==0.60.2
+maggma==0.62.1
     # via pyEQL (setup.py)
 markdown-it-py==3.0.0
     # via
     #   mdit-py-plugins
     #   myst-parser
     #   rich
-markupsafe==2.1.3
+markupsafe==2.1.5
     # via
     #   jinja2
     #   werkzeug
@@ -147,7 +147,7 @@ mongogrant==0.3.3
     # via maggma
 mongomock==4.1.2
     # via maggma
-monty==2023.11.3
+monty==2024.2.2
     # via
     #   maggma
     #   pyEQL (setup.py)
@@ -168,7 +168,7 @@ networkx==3.2.1
     # via pymatgen
 nodeenv==1.8.0
     # via pre-commit
-numpy==1.26.3
+numpy==1.26.4
     # via
     #   contourpy
     #   maggma
@@ -179,7 +179,7 @@ numpy==1.26.3
     #   pymatgen
     #   scipy
     #   spglib
-orjson==3.9.10
+orjson==3.9.13
     # via maggma
 packaging==23.2
     # via
@@ -192,7 +192,7 @@ packaging==23.2
     #   tox
 palettable==3.3.3
     # via pymatgen
-pandas==2.1.4
[…]
