+++ requirements/windows-latest_py3.9.txt
@@ -18,13 +18,13 @@ bcrypt==4.1.2
     # via paramiko
 blinker==1.7.0
     # via flask
-boto3==1.34.15
+boto3==1.34.35
     # via maggma
-botocore==1.34.15
+botocore==1.34.35
     # via
     #   boto3
     #   s3transfer
-certifi==2023.11.17
+certifi==2024.2.2
     # via requests
 cffi==1.16.0
     # via
@@ -43,21 +43,21 @@ colorama==0.4.6
     #   tqdm
 contourpy==1.2.0
     # via matplotlib
-cryptography==42.0.0
+cryptography==42.0.2
     # via paramiko
 cycler==0.12.1
     # via matplotlib
-dnspython==2.4.2
+dnspython==2.5.0
     # via
     #   maggma
     #   pymongo
 exceptiongroup==1.2.0
     # via anyio
-fastapi==0.109.1
+fastapi==0.109.2
     # via maggma
-flask==3.0.0
+flask==3.0.2
     # via mongogrant
-fonttools==4.47.0
+fonttools==4.48.1
     # via matplotlib
 future==0.18.3
     # via uncertainties
@@ -75,7 +75,7 @@ importlib-resources==6.1.1
     # via matplotlib
 itsdangerous==2.1.2
     # via flask
-jinja2==3.1.2
+jinja2==3.1.3
     # via flask
 jmespath==1.0.1
     # via
@@ -83,7 +83,7 @@ jmespath==1.0.1
     #   botocore
 joblib==1.3.2
     # via pymatgen
-jsonschema==4.20.0
+jsonschema==4.21.1
     # via maggma
 jsonschema-specifications==2023.12.1
     # via jsonschema
@@ -91,9 +91,9 @@ kiwisolver==1.4.5
     # via matplotlib
 latexcodec==2.0.1
     # via pybtex
-maggma==0.60.2
+maggma==0.62.1
     # via pyEQL (setup.py)
-markupsafe==2.1.3
+markupsafe==2.1.5
     # via
     #   jinja2
     #   werkzeug
@@ -103,7 +103,7 @@ mongogrant==0.3.3
     # via maggma
 mongomock==4.1.2
     # via maggma
-monty==2023.11.3
+monty==2024.2.2
     # via
     #   maggma
     #   pyEQL (setup.py)
@@ -114,7 +114,7 @@ msgpack==1.0.7
     # via maggma
 networkx==3.2.1
     # via pymatgen
-numpy==1.26.3
+numpy==1.26.4
     # via
     #   contourpy
     #   maggma
@@ -125,7 +125,7 @@ numpy==1.26.3
     #   pymatgen
     #   scipy
     #   spglib
-orjson==3.9.10
+orjson==3.9.13
     # via maggma
 packaging==23.2
     # via
@@ -134,7 +134,7 @@ packaging==23.2
     #   plotly
 palettable==3.3.3
     # via pymatgen
-pandas==2.1.4
+pandas==2.2.0
     # via pymatgen
 paramiko==3.4.0
     # via sshtunnel
@@ -152,18 +152,18 @@ pybtex==0.24.0
     # via pymatgen
 pycparser==2.21
     # via cffi
-pydantic==2.5.3
+pydantic==2.6.1
     # via
     #   fastapi
     #   maggma
     #   pydantic-settings
-pydantic-core==2.14.6
+pydantic-core==2.16.2
     # via pydantic
 pydantic-settings==2.1.0
     # via maggma
-pydash==7.0.6
+pydash==7.0.7
     # via maggma
-pymatgen==2023.12.18
+pymatgen==2024.1.27
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
[…]
