### 2024-02-15 c6e52e2
Solution: add default_diffusion_coeff
