+++ tests/test_solution.py
@@ -138,9 +138,12 @@ def test_diffusion_transport(s1, s2):
     assert np.isclose(s_dilute.get_transport_number("Cl-"), 0.604, atol=1e-3)
 
     # test setting a default value
+    s2.default_diffusion_coeff = 0
     assert s2.get_diffusion_coefficient("Cs+").magnitude == 0
-    assert s2.get_diffusion_coefficient("Cs+", default=1e-9, activity_correction=False).magnitude == 1e-9
-    assert s2.get_diffusion_coefficient("Cs+", default=1e-9, activity_correction=True).magnitude < 1e-9
+    s2.default_diffusion_coeff = 1e-9
+    assert s2.get_diffusion_coefficient("Cs+", activity_correction=False).magnitude == 1e-9
+    s2.default_diffusion_coeff = 0
+    assert s2.get_diffusion_coefficient("Cs+", activity_correction=True).magnitude < 1e-9
     d25 = s2.get_diffusion_coefficient("Na+", activity_correction=False).magnitude
     nu25 = s2.water_substance.nu
     s2.temperature = "40 degC"
