### 2024-02-15 be3a1a4
docstring cleanup
