### 2024-02-17 #6295 @orionarcher
Awesome work from @rkingsbury. I encountered no issues installing the package or following along with the examples. This work already appears to be widely used (49 stars) and the buildout of documentation and functionality is sure to make it more useful. I feel this work meets the standards of JOSS and should be accepted.
