### 2024-02-20 KingsburyLab/pyEQL#102 @yuxuanzhuang
JOSS Review--Documentation
This is part of the review process for https://github.com/openjournals/joss-reviews/issues/6295

Great work!! I've noticed a few minor issues that could be addressed:

- In https://pyeql.readthedocs.io/en/latest/, s1 = pyEQL.Solution -> `s1 = Solution()`.

- In https://pyeql.readthedocs.io/en/latest/chemistry.html#how-to-enter-valid-chemical-formulas:
   - The word "interpret" is misspelled as "inerpret."
   - The chemical formula for Sodium Sulfate should be`Na2SO4`.

- In https://pyeql.readthedocs.io/en/latest/mixing.html, there is an issue with an equation not being shown correctly.

- In https://pyeql.readthedocs.io/en/latest/class_solution.html and https://pyeql.readthedocs.io/en/latest/internal.html have rendering problems
   - Not all `Return` is displayed correctly.
   - Examples do not appear correctly.
   - Links in the `see also` sections are not functioning as expected.
