### 2024-02-21 9046971
Bump pymatgen from 2024.1.27 to 2024.2.20 in /requirements
Bumps [pymatgen](https://github.com/materialsproject/pymatgen) from 2024.1.27 to 2024.2.20.
- [Release notes](https://github.com/materialsproject/pymatgen/releases)
- [Changelog](https://github.com/materialsproject/pymatgen/blob/master/docs/CHANGES.md)
- [Commits](https://github.com/materialsproject/pymatgen/compare/v2024.1.27...v2024.2.20)

---
updated-dependencies:
- dependency-name: pymatgen
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
