+++ requirements/macos-latest_py3.11_extras.txt
@@ -229,7 +229,7 @@ pygments==2.17.2
     # via
     #   rich
     #   sphinx
-pymatgen==2024.1.27
+pymatgen==2024.2.20
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
