+++ requirements/macos-latest_py3.9_extras.txt
@@ -239,7 +239,7 @@ pygments==2.17.2
     # via
     #   rich
     #   sphinx
-pymatgen==2024.1.27
+pymatgen==2024.2.20
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
