+++ requirements/windows-latest_py3.10_extras.txt
@@ -240,7 +240,7 @@ pygments==2.17.2
     # via
     #   rich
     #   sphinx
-pymatgen==2024.1.27
+pymatgen==2024.2.20
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
