+++ requirements/windows-latest_py3.11_extras.txt
@@ -10,7 +10,7 @@ alabaster==0.7.16
     # via sphinx
 annotated-types==0.6.0
     # via pydantic
-anyio==4.2.0
+anyio==4.3.0
     # via starlette
 attrs==23.2.0
     # via
@@ -20,13 +20,13 @@ babel==2.14.0
     # via sphinx
 bcrypt==4.1.2
     # via paramiko
-black==24.1.1
+black==24.2.0
     # via pyEQL (setup.py)
 blinker==1.7.0
     # via flask
-boto3==1.34.35
+boto3==1.34.46
     # via maggma
-botocore==1.34.35
+botocore==1.34.46
     # via
     #   boto3
     #   s3transfer
@@ -55,17 +55,15 @@ colorama==0.4.6
     #   tqdm
 contourpy==1.2.0
     # via matplotlib
-coverage[toml]==7.4.1
-    # via
-    #   coverage
-    #   pytest-cov
-cryptography==42.0.2
+coverage[toml]==7.4.2
+    # via pytest-cov
+cryptography==42.0.4
     # via paramiko
 cycler==0.12.1
     # via matplotlib
 distlib==0.3.8
     # via virtualenv
-dnspython==2.5.0
+dnspython==2.6.1
     # via
     #   maggma
     #   pymongo
@@ -82,15 +80,15 @@ filelock==3.13.1
     #   virtualenv
 flask==3.0.2
     # via mongogrant
-fonttools==4.48.1
+fonttools==4.49.0
     # via matplotlib
-future==0.18.3
+future==1.0.0
     # via uncertainties
 h11==0.14.0
     # via uvicorn
 iapws==1.5.3
     # via pyEQL (setup.py)
-identify==2.5.33
+identify==2.5.35
     # via pre-commit
 idna==3.6
     # via
@@ -123,7 +121,7 @@ latexcodec==2.0.1
     # via pybtex
 linkify-it-py==2.0.3
     # via myst-parser
-maggma==0.62.1
+maggma==0.63.2
     # via pyEQL (setup.py)
 markdown-it-py==3.0.0
     # via
@@ -134,7 +132,7 @@ markupsafe==2.1.5
     # via
     #   jinja2
     #   werkzeug
-matplotlib==3.8.2
+matplotlib==3.8.3
     # via pymatgen
 mdit-py-plugins==0.4.0
     # via myst-parser
@@ -176,7 +174,7 @@ numpy==1.26.4
     #   pymatgen
     #   scipy
     #   spglib
-orjson==3.9.13
+orjson==3.9.14
     # via maggma
 packaging==23.2
     # via
@@ -207,13 +205,13 @@ platformdirs==4.2.0
     # via
     #   black
     #   virtualenv
-plotly==5.18.0
+plotly==5.19.0
     # via pymatgen
 pluggy==1.4.0
     # via
     #   pytest
     #   tox
-pre-commit==3.6.0
+pre-commit==3.6.2
     # via pyEQL (setup.py)
 py==1.11.0
     # via tox
@@ -228,7 +226,7 @@ pydantic==2.6.1
     #   pydantic-settings
 pydantic-core==2.16.2
     # via pydantic
-pydantic-settings==2.1.0
+pydantic-settings==2.2.1
     # via maggma
 pydash==7.0.7
     # via maggma
@@ -236,7 +234,7 @@ pygments==2.17.2
     # via
     #   rich
     #   sphinx
-pymatgen==2024.1.27
+pymatgen==2024.2.20
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
@@ -248,7 +246,7 @@ pyparsing==3.1.1
[…]
