+++ requirements/windows-latest_py3.11.txt
@@ -8,7 +8,7 @@ aioitertools==0.11.0
     # via maggma
 annotated-types==0.6.0
     # via pydantic
-anyio==4.2.0
+anyio==4.3.0
     # via starlette
 attrs==23.2.0
     # via
@@ -18,9 +18,9 @@ bcrypt==4.1.2
     # via paramiko
 blinker==1.7.0
     # via flask
-boto3==1.34.35
+boto3==1.34.46
     # via maggma
-botocore==1.34.35
+botocore==1.34.46
     # via
     #   boto3
     #   s3transfer
@@ -43,11 +43,11 @@ colorama==0.4.6
     #   tqdm
 contourpy==1.2.0
     # via matplotlib
-cryptography==42.0.2
+cryptography==42.0.4
     # via paramiko
 cycler==0.12.1
     # via matplotlib
-dnspython==2.5.0
+dnspython==2.6.1
     # via
     #   maggma
     #   pymongo
@@ -55,9 +55,9 @@ fastapi==0.109.2
     # via maggma
 flask==3.0.2
     # via mongogrant
-fonttools==4.48.1
+fonttools==4.49.0
     # via matplotlib
-future==0.18.3
+future==1.0.0
     # via uncertainties
 h11==0.14.0
     # via uvicorn
@@ -85,13 +85,13 @@ kiwisolver==1.4.5
     # via matplotlib
 latexcodec==2.0.1
     # via pybtex
-maggma==0.62.1
+maggma==0.63.2
     # via pyEQL (setup.py)
 markupsafe==2.1.5
     # via
     #   jinja2
     #   werkzeug
-matplotlib==3.8.2
+matplotlib==3.8.3
     # via pymatgen
 mongogrant==0.3.3
     # via maggma
@@ -119,7 +119,7 @@ numpy==1.26.4
     #   pymatgen
     #   scipy
     #   spglib
-orjson==3.9.13
+orjson==3.9.14
     # via maggma
 packaging==23.2
     # via
@@ -140,7 +140,7 @@ pillow==10.2.0
     # via matplotlib
 pint==0.23
     # via pyEQL (setup.py)
-plotly==5.18.0
+plotly==5.19.0
     # via pymatgen
 pybtex==0.24.0
     # via pymatgen
@@ -153,11 +153,11 @@ pydantic==2.6.1
     #   pydantic-settings
 pydantic-core==2.16.2
     # via pydantic
-pydantic-settings==2.1.0
+pydantic-settings==2.2.1
     # via maggma
 pydash==7.0.7
     # via maggma
-pymatgen==2024.1.27
+pymatgen==2024.2.20
     # via pyEQL (setup.py)
 pymongo==4.6.1
     # via
@@ -191,7 +191,7 @@ requests==2.31.0
     # via
     #   mongogrant
     #   pymatgen
-rpds-py==0.17.1
+rpds-py==0.18.0
     # via
     #   jsonschema
     #   referencing
@@ -218,7 +218,7 @@ six==1.16.0
     #   python-dateutil
 sniffio==1.3.0
     # via anyio
-spglib==2.3.0
+spglib==2.3.1
     # via pymatgen
 sshtunnel==0.4.0
     # via maggma
@@ -230,7 +230,7 @@ tabulate==0.9.0
     # via pymatgen
 tenacity==8.2.3
     # via plotly
-tqdm==4.66.1
+tqdm==4.66.2
     # via
     #   maggma
     #   pymatgen
@@ -241,7 +241,7 @@ typing-extensions==4.9.0
     #   pydantic
     #   pydantic-core
     #   pydash
-tzdata==2023.4
+tzdata==2024.1
     # via pandas
[…]
