### 2024-02-25 9209926
Solution.get_total_amount: bufix for mass based units
