+++ src/pyEQL/solute.py
@@ -115,13 +115,11 @@ class Solute:
         mw = f"{float(pmg_ion.weight / factor)} g/mol"  # weight is a FloatWithUnit
         chemsys = pmg_ion.chemical_system
         # store only the most likely oxi_state guesses
-        oxi_states = pmg_ion.oxi_state_guesses(all_oxi_states=True)
-        # TODO - hack to work around a pymatgen bug in Composition
-        # https://github.com/materialsproject/pymatgen/issues/3324
-        if oxi_states == []:
-            oxi_states = {els[0]: 0.0} if rform in ["O2(aq)", "O3(aq)", "Cl2(aq)", "F2(aq)"] else {}
-        else:
-            oxi_states = oxi_states[0]
+        try:
+            oxi_states = pmg_ion.oxi_state_guesses(all_oxi_states=True)
+            oxi_states == {} if oxi_states == () else oxi_states[0]
+        except (IndexError, ValueError):
+            oxi_states = {}
 
         return cls(
             rform,
