+++ CHANGELOG.md
@@ -5,10 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.12.1] - 2024-02-25
 
 ### Fixed
 
+- `Solute.from_formula`: Fix bug in which an uncaught exception could occur when
+  if `pymatgen` failed to guess the oxidation state of a solute. (Issue #103 - thanks to @xiaoxiaozhu123 for reporting).
 - `Solution.get_total_amount`: Bugfix that would cause the method to fail if
   mass-based units such as mg/L or ppm were requested.
 
