### 2024-02-25 5f4a79e
Solution.get_total_amount: bugfix for aliased units like ppm
