### 2024-02-26 #6295 @JacksonBurns
Installation and testing was seamless. I was able to follow the examples without issues. I tried entering some non-physical solutions and some ions not in the database and was met with reasonable warnings, which I appreciate.

I will add that the review of 'competing' packages was especially thorough! Well done. Paper overall is concise and convincing, and the software seems outstanding in general.

There are a few small errors I stumbled across that should be fixed, but after that this package looks ready!

List of some small things:
 - the ending copyright year on the license (2023) is now outdated
 - I think you are missing an `@` in your paper on this line: https://github.com/KingsburyLab/pyEQL/blob/fd35750721f4c013091cab2c14fee1a5934dc1be/docs/joss/paper.md?plain=1#L66
 - There is an extra `:::` on this line of the docs: https://github.com/KingsburyLab/pyEQL/blob/c43b2e6d97b3cfa7c023fe75405e85c9dbbbed67/docs/contributing.md?plain=1#L11
 - Might be missing a hyperlink on this line of the docs for `[Sphinx]` (?): https://github.com/KingsburyLab/pyEQL/blob/c43b2e6d97b3cfa7c023fe75405e85c9dbbbed67/docs/contributing.md?plain=1#L20
 - The sections of code and text in the contributing page are flipped from this line onwards: https://github.com/KingsburyLab/pyEQL/blob/c43b2e6d97b3cfa7c023fe75405e85c9dbbbed67/docs/contributing.md?plain=1#L47 so it renders like this:
![image](https://github.com/openjournals/joss-reviews/assets/33505528/a628f1cb-bd5d-4921-ab0d-70252f8a5d8d)

One larger idea, which is out of scope of this review but I thought I might share in case it is helpful (feel free to ignore):

I am sure you are aware of this because of your very thorough requirements directory, but the dependencies are pretty expansive (`pip install pyEQL` in a bare python 3.9 conda environment installed 94 packages for me). I generated this dependency graph:
<summary> tree </summary>

```
(pyeql_39) jackson@jackson-Precision-7540:~$ pipdeptree
pip==24.0
pipdeptree==2.15.1
pyEQL==0.12.2
├── iapws [required: Any, installed: 1.5.3]
│   └── scipy [required: >=1.2, installed: 1.12.0]
│       └── numpy [required: >=1.22.4,<1.29.0, installed: 1.26.4]
├── maggma [required: >=0.57.5, installed: 0.63.3]
│   ├── aioitertools [required: >=0.5.1, installed: 0.11.0]
│   │   └── typing-extensions [required: >=4.0, installed: 4.10.0]
│   ├── boto3 [required: >=1.20.41, installed: 1.34.49]
│   │   ├── botocore [required: >=1.34.49,<1.35.0, installed: 1.34.49]
│   │   │   ├── jmespath [required: >=0.7.1,<2.0.0, installed: 1.0.1]
│   │   │   ├── python-dateutil [required: >=2.1,<3.0.0, installed: 2.8.2]
│   │   │   │   └── six [required: >=1.5, installed: 1.16.0]
│   │   │   └── urllib3 [required: >=1.25.4,<1.27, installed: 1.26.18]
│   │   ├── jmespath [required: >=0.7.1,<2.0.0, installed: 1.0.1]
│   │   └── s3transfer [required: >=0.10.0,<0.11.0, installed: 0.10.0]
│   │       └── botocore [required: >=1.33.2,<2.0a.0, installed: 1.34.49]
│   │           ├── jmespath [required: >=0.7.1,<2.0.0, installed: 1.0.1]
│   │           ├── python-dateutil [required: >=2.1,<3.0.0, installed: 2.8.2]
│   │           │   └── six [required: >=1.5, installed: 1.16.0]
│   │           └── urllib3 [required: >=1.25.4,<1.27, installed: 1.26.18]
│   ├── dnspython [required: >=1.16.0, installed: 2.6.1]
│   ├── fastapi [required: >=0.42.0, installed: 0.110.0]
│   │   ├── pydantic [required: >=1.7.4,<3.0.0,!=2.1.0,!=2.0.1,!=2.0.0,!=1.8.1,!=1.8, installed: 2.6.2]
│   │   │   ├── annotated-types [required: >=0.4.0, installed: 0.6.0]
│   │   │   ├── pydantic-core [required: ==2.16.3, installed: 2.16.3]
│   │   │   │   └── typing-extensions [required: >=4.6.0,!=4.7.0, installed: 4.10.0]
│   │   │   └── typing-extensions [required: >=4.6.1, installed: 4.10.0]
│   │   ├── starlette [required: >=0.36.3,<0.37.0, installed: 0.36.3]
│   │   │   ├── anyio [required: >=3.4.0,<5, installed: 4.3.0]
│   │   │   │   ├── exceptiongroup [required: >=1.0.2, installed: 1.2.0]
│   │   │   │   ├── idna [required: >=2.8, installed: 3.6]
│   │   │   │   ├── sniffio [required: >=1.1, installed: 1.3.1]
│   │   │   │   └── typing-extensions [required: >=4.1, installed: 4.10.0]
│   │   │   └── typing-extensions [required: >=3.10.0, installed: 4.10.0]
│   │   └── typing-extensions [required: >=4.8.0, installed: 4.10.0]
│   ├── jsonschema [required: >=3.1.1, installed: 4.21.1]
│   │   ├── attrs [required: >=22.2.0, installed: 23.2.0]
│   │   ├── jsonschema-specifications [required: >=2023.03.6, installed: 2023.12.1]
│   │   │   └── referencing [required: >=0.31.0, installed: 0.33.0]
│   │   │       ├── attrs [required: >=22.2.0, installed: 23.2.0]
│   │   │       └── rpds-py [required: >=0.7.0, installed: 0.18.0]
│   │   ├── referencing [required: >=0.28.4, installed: 0.33.0]
│   │   │   ├── attrs [required: >=22.2.0, installed: 23.2.0]
│   │   │   └── rpds-py [required: >=0.7.0, installed: 0.18.0]
│   │   └── rpds-py [required: >=0.7.1, installed: 0.18.0]
│   ├── mongogrant [required: >=0.3.1, installed: 0.3.3]
│   │   ├── click [required: Any, installed: 8.1.7]
│   │   ├── flask [required: >=1.0, installed: 3.0.2]
│   │   │   ├── blinker [required: >=1.6.2, installed: 1.7.0]
│   │   │   ├── click [required: >=8.1.3, installed: 8.1.7]
│   │   │   ├── importlib-metadata [required: >=3.6.0, installed: 7.0.1]
│   │   │   │   └── zipp [required: >=0.5, installed: 3.17.0]
│   │   │   ├── itsdangerous [required: >=2.1.2, installed: 2.1.2]
│   │   │   ├── Jinja2 [required: >=3.1.2, installed: 3.1.3]
│   │   │   │   └── MarkupSafe [required: >=2.0, installed: 2.1.5]
│   │   │   └── werkzeug [required: >=3.0.0, installed: 3.0.1]
│   │   │       └── MarkupSafe [required: >=2.1.1, installed: 2.1.5]
│   │   ├── pymongo [required: >=3.8, installed: 4.6.2]
│   │   │   └── dnspython [required: >=1.16.0,<3.0.0, installed: 2.6.1]
│   │   └── requests [required: Any, installed: 2.31.0]
│   │       ├── certifi [required: >=2017.4.17, installed: 2024.2.2]
│   │       ├── charset-normalizer [required: >=2,<4, installed: 3.3.2]
│   │       ├── idna [required: >=2.5,<4, installed: 3.6]
│   │       └── urllib3 [required: >=1.21.1,<3, installed: 1.26.18]
│   ├── mongomock [required: >=3.10.0, installed: 4.1.2]
│   │   ├── packaging [required: Any, installed: 23.2]
│   │   └── sentinels [required: Any, installed: 1.0.0]
│   ├── monty [required: >=2023.9.25, installed: 2024.2.26]
│   ├── msgpack [required: >=0.5.6, installed: 1.0.7]
│   ├── numpy [required: >=1.17.3, installed: 1.26.4]
│   ├── orjson [required: >=3.9.0, installed: 3.9.15]
│   ├── pydantic [required: >=2.0, installed: 2.6.2]
│   │   ├── annotated-types [required: >=0.4.0, installed: 0.6.0]
│   │   ├── pydantic-core [required: ==2.16.3, installed: 2.16.3]
│   │   │   └── typing-extensions [required: >=4.6.0,!=4.7.0, installed: 4.10.0]
│   │   └── typing-extensions [required: >=4.6.1, installed: 4.10.0]
│   ├── pydantic-settings [required: >=2.0.3, installed: 2.2.1]
│   │   ├── pydantic [required: >=2.3.0, installed: 2.6.2]
│   │   │   ├── annotated-types [required: >=0.4.0, installed: 0.6.0]
│   │   │   ├── pydantic-core [required: ==2.16.3, installed: 2.16.3]
│   │   │   │   └── typing-extensions [required: >=4.6.0,!=4.7.0, installed: 4.10.0]
│   │   │   └── typing-extensions [required: >=4.6.1, installed: 4.10.0]
│   │   └── python-dotenv [required: >=0.21.0, installed: 1.0.1]
│   ├── pydash [required: >=4.1.0, installed: 7.0.7]
│   │   └── typing-extensions [required: >=3.10,!=4.6.0, installed: 4.10.0]
│   ├── pymongo [required: >=4.2.0, installed: 4.6.2]
│   │   └── dnspython [required: >=1.16.0,<3.0.0, installed: 2.6.1]
│   ├── python-dateutil [required: >=2.8.2, installed: 2.8.2]
│   │   └── six [required: >=1.5, installed: 1.16.0]
│   ├── pyzmq [required: >=24.0.1, installed: 25.1.2]
│   ├── ruamel.yaml [required: <0.18, installed: 0.17.40]
│   │   └── ruamel.yaml.clib [required: >=0.2.7, installed: 0.2.8]
│   ├── setuptools [required: Any, installed: 69.1.1]
│   ├── sshtunnel [required: >=0.1.5, installed: 0.4.0]
│   │   └── paramiko [required: >=2.7.2, installed: 3.4.0]
│   │       ├── bcrypt [required: >=3.2, installed: 4.1.2]
│   │       ├── cryptography [required: >=3.3, installed: 42.0.5]
│   │       │   └── cffi [required: >=1.12, installed: 1.16.0]
│   │       │       └── pycparser [required: Any, installed: 2.21]
│   │       └── PyNaCl [required: >=1.5, installed: 1.5.0]
│   │           └── cffi [required: >=1.4.1, installed: 1.16.0]
│   │               └── pycparser [required: Any, installed: 2.21]
│   ├── tqdm [required: >=4.19.6, installed: 4.66.2]
│   └── uvicorn [required: >=0.18.3, installed: 0.27.1]
│       ├── click [required: >=7.0, installed: 8.1.7]
│       ├── h11 [required: >=0.8, installed: 0.14.0]
│       └── typing-extensions [required: >=4.0, installed: 4.10.0]
├── monty [required: Any, installed: 2024.2.26]
├── numpy [required: Any, installed: 1.26.4]
├── phreeqpython [required: Any, installed: 1.5.0]
│   ├── periodictable [required: Any, installed: 1.6.1]
│   │   ├── numpy [required: Any, installed: 1.26.4]
│   │   └── pyparsing [required: Any, installed: 3.1.1]
│   └── scipy [required: Any, installed: 1.12.0]
│       └── numpy [required: >=1.22.4,<1.29.0, installed: 1.26.4]
├── Pint [required: >=0.19, installed: 0.23]
│   └── typing-extensions [required: Any, installed: 4.10.0]
├── pymatgen [required: >=2023.10.11, installed: 2024.2.23]
│   ├── joblib [required: Any, installed: 1.3.2]
│   ├── matplotlib [required: >=1.5, installed: 3.8.3]
│   │   ├── contourpy [required: >=1.0.1, installed: 1.2.0]
│   │   │   └── numpy [required: >=1.20,<2.0, installed: 1.26.4]
│   │   ├── cycler [required: >=0.10, installed: 0.12.1]
│   │   ├── fonttools [required: >=4.22.0, installed: 4.49.0]
│   │   ├── importlib-resources [required: >=3.2.0, installed: 6.1.2]
│   │   │   └── zipp [required: >=3.1.0, installed: 3.17.0]
│   │   ├── kiwisolver [required: >=1.3.1, installed: 1.4.5]
│   │   ├── numpy [required: >=1.21,<2, installed: 1.26.4]
│   │   ├── packaging [required: >=20.0, installed: 23.2]
│   │   ├── pillow [required: >=8, installed: 10.2.0]
│   │   ├── pyparsing [required: >=2.3.1, installed: 3.1.1]
│   │   └── python-dateutil [required: >=2.7, installed: 2.8.2]
│   │       └── six [required: >=1.5, installed: 1.16.0]
│   ├── monty [required: >=2024.2.2, installed: 2024.2.26]
│   ├── networkx [required: >=2.2, installed: 3.2.1]
│   ├── numpy [required: >=1.25.0, installed: 1.26.4]
│   ├── palettable [required: >=3.1.1, installed: 3.3.3]
│   ├── pandas [required: Any, installed: 2.2.1]
│   │   ├── numpy [required: >=1.22.4,<2, installed: 1.26.4]
│   │   ├── python-dateutil [required: >=2.8.2, installed: 2.8.2]
│   │   │   └── six [required: >=1.5, installed: 1.16.0]
│   │   ├── pytz [required: >=2020.1, installed: 2024.1]
│   │   └── tzdata [required: >=2022.7, installed: 2024.1]
│   ├── plotly [required: >=4.5.0, installed: 5.19.0]
│   │   ├── packaging [required: Any, installed: 23.2]
│   │   └── tenacity [required: >=6.2.0, installed: 8.2.3]
│   ├── pybtex [required: Any, installed: 0.24.0]
│   │   ├── latexcodec [required: >=1.0.4, installed: 2.0.1]
│   │   │   └── six [required: >=1.4.1, installed: 1.16.0]
│   │   ├── PyYAML [required: >=3.01, installed: 6.0.1]
│   │   └── six [required: Any, installed: 1.16.0]
│   ├── requests [required: Any, installed: 2.31.0]
│   │   ├── certifi [required: >=2017.4.17, installed: 2024.2.2]
│   │   ├── charset-normalizer [required: >=2,<4, installed: 3.3.2]
│   │   ├── idna [required: >=2.5,<4, installed: 3.6]
│   │   └── urllib3 [required: >=1.21.1,<3, installed: 1.26.18]
│   ├── ruamel.yaml [required: >=0.17.0, installed: 0.17.40]
│   │   └── ruamel.yaml.clib [required: >=0.2.7, installed: 0.2.8]
│   ├── scipy [required: >=1.5.0, installed: 1.12.0]
│   │   └── numpy [required: >=1.22.4,<1.29.0, installed: 1.26.4]
│   ├── spglib [required: >=2.0.2, installed: 2.3.1]
│   │   └── numpy [required: Any, installed: 1.26.4]
│   ├── sympy [required: Any, installed: 1.12]
│   │   └── mpmath [required: >=0.19, installed: 1.3.0]
│   ├── tabulate [required: Any, installed: 0.9.0]
│   ├── tqdm [required: Any, installed: 4.66.2]
│   └── uncertainties [required: >=3.1.4, installed: 3.1.7]
│       └── future [required: Any, installed: 1.0.0]
└── scipy [required: Any, installed: 1.12.0]
    └── numpy [required: >=1.22.4,<1.29.0, installed: 1.26.4]
wheel==0.42.0
```

and it seems that most of the bulk is coming from `maggma` and `pymatgen`. If most of your end users are within the materials project world I don't see this being a problem, but it may impact reuse potential for those trying to extend `pyEQL` into other domains. If possible, could these be removed or made optional (to maintain optional perfect compatibility with materials project)? A quick [search of the `pyEQL` repo](https://github.com/search?q=repo%3AKingsburyLab%2FpyEQL+pymatgen+language%3APython&type=code&l=Python) shows that `pymatgen` (as also described in the README) is only used to standardize some input data - could this be achieved with another (smaller) package? Similarly for `maggma`, [search](https://github.com/search?q=repo%3AKingsburyLab%2FpyEQL+pymatgen+language%3APython&type=code&l=Python) (and README) show it used only for serializing and de-serealizing - could you wrap `pymongo` yourself and remove `maggma`?
