+++ requirements/windows-latest_py3.9_extras.txt
@@ -184,7 +184,7 @@ numpy==1.26.4
     #   pymatgen
     #   scipy
     #   spglib
-orjson==3.9.14
+orjson==3.9.15
     # via maggma
 packaging==23.2
     # via
