### 2024-02-27 KingsburyLab/pyEQL#102 @rkingsbury
Reproducing here some additional comments from the JOSS review related to Docs cleanups:

List of some small things:

   -  the ending copyright year on the license (2023) is now outdated
   -  I think you are missing an @ in your paper on this line: https://github.com/KingsburyLab/pyEQL/blob/fd35750721f4c013091cab2c14fee1a5934dc1be/docs/joss/paper.md?plain=1#L66
    - There is an extra ::: on this line of the docs: https://github.com/KingsburyLab/pyEQL/blob/c43b2e6d97b3cfa7c023fe75405e85c9dbbbed67/docs/contributing.md?plain=1#L11
    - Might be missing a hyperlink on this line of the docs for [Sphinx] (?): https://github.com/KingsburyLab/pyEQL/blob/c43b2e6d97b3cfa7c023fe75405e85c9dbbbed67/docs/contributing.md?plain=1#L20
    - The sections of code and text in the contributing page are flipped from this line onwards: https://github.com/KingsburyLab/pyEQL/blob/c43b2e6d97b3cfa7c023fe75405e85c9dbbbed67/docs/contributing.md?plain=1#L47 so it renders like this:
