### 2024-02-27 #6295 @rkingsbury
> Installation and testing was seamless. I was able to follow the examples without issues. I tried entering some non-physical solutions and some ions not in the database and was met with reasonable warnings, which I appreciate.
>
> I will add that the review of 'competing' packages was especially thorough! Well done. Paper overall is concise and convincing, and the software seems outstanding in general.

Thank you very much @JacksonBurns ! I appreciate your thorough testing and identifying the documentation issues.

> One larger idea, which is out of scope of this review but I thought I might share in case it is helpful (feel free to ignore):
>
> I am sure you are aware of this because of your very thorough requirements directory, but the dependencies are pretty expansive (`pip install pyEQL` in a bare python 3.9 conda environment installed 94 packages for me). I generated this dependency graph:
> tree

Great suggestions. I have thought about this in the abstract but I did not realize the extent of the dependency bloat until seeing the tree you generated (aside: I was unaware of `pipdeptree`, so thanks for that!). I am a maintainer of `maggma` and I think we should explore more optional dependency groups to manage that, because many of those are not needed for basic mongo-style querying. Or, as you suggest, for the specific ways we use `maggma` in `pyEQL`, we might be able to get away with a custom solution.

You are correct that the main purpose of having `pymatgen` in here is so that formulas can be standardized in a way that `pymatgen` will understand, and also to leverage `pymatgen` for chemical informatics such as molecular weights, oxidation states, etc. based on those formulas. There are alternatives, but I know that efforts are underway to make `pymatgen` itself leaner so I think it's likely I will stick with it in `pyEQL` for the near future.
