### 2024-02-29 #6295 @yuxuanzhuang
I love the features of this package and even plan to add it to my research toolbox. I've started a separate issue about documentation issues in the repository, which the author is already addressing. Aside from that, the software is well-documented and the tests are abundant.

I do have an additional comment/question: It appears that the package does not support Mac computers equipped with Apple silicon (although it can be installed without any issue). This seems to be due to the `phreeqpython` package lacks support for it:

```python
> s1 = Solution()
OSError: dlopen(/xxx/lib/python3.10/site-packages/phreeqpython/./lib/viphreeqc.dylib, 0x0006): tried: '/xxx/lib/python3.10/site-packages/phreeqpython/./lib/viphreeqc.dylib' (mach-o file, but is an incompatible architecture (have 'x86_64', need 'arm64')),
```

Perhaps it's better to document this compatibility issue somewhere?

I would definitely recommend it to be accepted. Great work!!!
