### 2024-02-29 KingsburyLab/pyEQL#109 @rkingsbury
Incomplete support for Apple silicon (M1, M2, etc.)
~~As reported by  @yuxuanzhuang during the [JOSS review](https://github.com/openjournals/joss-reviews/issues/6295#issuecomment-1970472354), `phreeqptyon` does not work on Apple silicon, and thereby causes `Solution` to fail (even though `pyEQL` can be installed):~~

> s1 = Solution()
> OSError: dlopen(/xxx/lib/python3.10/site-packages/phreeqpython/./lib/viphreeqc.dylib, 0x0006): tried: '/xxx/lib/python3.10/site-packages/phreeqpython/./lib/viphreeqc.dylib' (mach-o file, but is an incompatible architecture (have 'x86_64', need 'arm64')),

~~We need to work around this. At a minimum, it should be possible to instantiate and use a `Solution` without the `equilibrate` function (which is the only thing that relies on `phreeqpython` by default).~~

**UPDATE as of v0.14.0**

`pyEQL` currently works on Apple Silicon, but any features that depend on PHREEQC (e.g., the `equilibrate` method) will raise errors. Changes have been made to the documentation to reflect this, and an Error is logged when you create a `Solution` on Apple silicon notifying you of the limitation.

It would be great to have full support for Apple silicon; however, it isn't clear to me whether upstream PHREEQC / `iphreeqc` / `phreeqpy` support Apple Silicon. If they do, this issue will add motivation to address #61 so that support is not dependent on `phreeqpython` (as opposed to upstream `iphreeqc`).
