+++ .github/workflows/release.yml
@@ -38,5 +38,5 @@ jobs:
         tox -e clean,build
 
     - name: Upload to PyPi
-      uses: pypa/gh-action-pypi-publish@v1.8.11
+      uses: pypa/gh-action-pypi-publish@v1.8.12
 
