### 2024-03-04 b4649e6
CI: add Apple M1 runner to github actions
