+++ tests/test_phreeqc.py
@@ -7,12 +7,16 @@ used by pyEQL's Solution class
 """
 
 import logging
+import platform
 
 import numpy as np
 import pytest
 
 from pyEQL import Solution
 from pyEQL.engines import PhreeqcEOS
+
+if platform.machine().startswith("arm64") and platform.system().startswith("Darwin"):
+    pytest.skip("skipping PHREEQC tests because arm64 is not supported", allow_module_level=True)
 
 
 @pytest.fixture()
