### 2024-03-04 KingsburyLab/pyEQL#109 @rkingsbury
@yuxuanzhuang can you provide a more complete code snippet of the steps that led to this error? I added the `macos-14` runner to my CI (which is supposed to be based on an M1 chip) and all tests pass, so I'm a bit puzzled. Thanks!
