+++ src/pyEQL/equilibrium.py
@@ -4,7 +4,7 @@ redox, complexation, etc.).
 
 NOTE: these methods are not currently used but are here for the future.
 
-:copyright: 2013-2023 by Ryan S. Kingsbury
+:copyright: 2013-2024 by Ryan S. Kingsbury
 :license: LGPL, see LICENSE for more details.
 
 """
