+++ docs/installation.md
@@ -6,10 +6,9 @@
 
 We highly recommend installing python in an isolated environment using [`conda`](https://docs.conda.io/en/latest/) (or its speedier, backward-compatible successor, [mamba](https://mamba.readthedocs.io/en/latest/)). In particular, we recommend the [miniforge](https://github.com/conda-forge/miniforge#miniforge3) or [mambaforge](https://github.com/conda-forge/miniforge#mambaforge) distributions of Python, which are lightweight distributions of conda that automatically activate the `conda-forge` channel for up-to-date scientific packages.
 
-
-:::{note}
+```{note}
 If you are on a Windows machine, we recommend you install the [Windows Subsystem for Linux (WSL)](https://learn.microsoft.com/en-us/windows/wsl/install) and set up your conda environments inside the WSL environment.
-:::
+```
 
 After installing `conda` / `mamba`, follow their instructions to create an environment. The steps should be similar to the following:
 
@@ -28,25 +27,29 @@ pip install pyEQL
 
 This should automatically pull in the required [dependencies](#other-dependencies) as well.
 
-:::{important}
+```{important}
 If you are NOT using a `conda` environment, may have to run 'pip3' rather than 'pip'. This will be the case if Python 2.x and Python 3.x are installed side-by-side on your system.
 You can tell if this is the case by typing the following command:
 
 ```
+
 $ python --version
 Python 2.7.12
+
 ```
 
 This means Python 2.x is installed. If you run 'pip install' it will point to the Python 2.7 installation, but pyEQL
 only works on Python 3. So, try this:
 
 ```
+
 $ python3 --version
 Python 3.9.7
+
 ```
 
 To get to Python 3.x, you have to type 'python3'. In this case, you would run 'pip3 install'
-:::
+```
 
 ## Other dependencies
 
@@ -61,7 +64,6 @@ pyEQL also requires the following packages:
 - [numpy](http://numpy.org/)
 
 If you use pip to install pyEQL (recommended), they should be installed automatically.
-
 
 ## Installing the development branch
 
@@ -85,6 +87,6 @@ Then install by executing:
 pip install -e pyEQL
 ```
 
-:::{note}
+```{note}
 You may have to run 'pip3' rather than 'pip'. See the note in the [pip install](#pip-install) section.
-:::
+```

[…]
