+++ src/pyEQL/activity_correction.py
@@ -37,7 +37,9 @@ def _debye_parameter_B(temperature: str = "25 degC") -> Quantity:
     Notes:
         The parameter B is equal to:
 
-        .. math:: B = ( \\frac{{2 N_A \\rho_w e^2}{\\epsilon_o \\epsilon_r k T}) ^ {1 \\over 2}
+        .. math::
+
+            B = \bigg( \frac{2 N_A \rho_w e^2}{\epsilon_o \epsilon_r k T} \bigg) ^ {\frac{1}{2}}
 
     References:
         Bockris and Reddy. /Modern Electrochemistry/, vol 1. Plenum/Rosetta, 1977, p.210.
@@ -66,7 +68,7 @@ def _debye_parameter_B(temperature: str = "25 degC") -> Quantity:
 
 
 def _debye_parameter_activity(temperature: str = "25 degC") -> "Quantity":
-    """
+    r"""
     Return the constant A for use in the Debye-Huckel limiting law (base e).
 
     Args:
@@ -82,7 +84,8 @@ def _debye_parameter_activity(temperature: str = "25 degC") -> "Quantity":
         The parameter A is equal to:
 
         ..  math::
-            A^{\\gamma} = {e^3 ( 2 \\pi N_A {\\rho})^{0.5} \\over (4 \\pi \\epsilon_o \\epsilon_r k T)^{1.5}}
+
+            A^{\gamma} = \frac{e^3 \big( 2 \pi N_A \rho \big)^{0.5}}{(4 \pi \epsilon_o \epsilon_r k T)^{1.5}}
 
         Note that this equation returns the parameter value that can be used to calculate
         the natural logarithm of the activity coefficient. For base 10, divide the
@@ -97,7 +100,7 @@ def _debye_parameter_activity(temperature: str = "25 degC") -> "Quantity":
         https://en.wikipedia.org/wiki/Debye%E2%80%93H%C3%BCckel_equation
 
     See Also:
-        :py:func:`_debye_parameter_osmotic`
+        :func:`_debye_parameter_osmotic`
 
     """
     T = ureg.Quantity(temperature)
@@ -112,7 +115,7 @@ def _debye_parameter_activity(temperature: str = "25 degC") -> "Quantity":
         / (4 * math.pi * ureg.epsilon_0 * water_substance.epsilon * ureg.boltzmann_constant * T) ** 1.5
     )
 
-    logger.info(f"Computed Debye-Huckel Limiting Law Constant A^{{\\gamma}} = {debyeparam} at {temperature}")
+    logger.info(rf"Computed Debye-Huckel Limiting Law Constant A^{{\gamma}} = {debyeparam} at {temperature}")
     return debyeparam.to("kg ** 0.5 / mol ** 0.5")
 
 
@@ -120,38 +123,33 @@ def _debye_parameter_osmotic(temperature="25 degC"):
     r"""
     Return the constant A_phi for use in calculating the osmotic coefficient according to Debye-Huckel theory.
 
-    Parameters
-    ----------
-    temperature : str Quantity, optional
-                  String representing the temperature of the solution. Defaults to '25 degC' if not specified.
+    Args:
+        temperature: String representing the temperature of the solution. Defaults to '25 degC' if not specified.
 
-    Notes
-    -----
-    Not to be confused with the Debye-Huckel constant used for activity coefficients in the limiting law.
-    Takes the value 0.392 at 25 C.
-    This constant is calculated according to: [kim]_ [arch]_
+    Notes:
+        Not to be confused with the Debye-Huckel constant used for activity coefficients in the limiting law.
+        Takes the value 0.392 at 25 C.
[…]
