### 2024-03-05 3cd6b3f
Solution.from_file: Docs update and bugfix
