### 2024-03-05 43bd634
utils: FormulaDict fix for python 3.12
