### 2024-03-05 0012e68
NativeEOS: try/except for OSError (Apple M1)
