### 2024-03-05 e49a12f
CI: separate more comprehensive testing for release
