+++ .github/workflows/post-process.yml
@@ -1,19 +1,70 @@
-name: Post-process
+name: Post-merge
 
+# triggered when PRs are merged into main branch
 on:
-  workflow_run:
+  pull_request:
     branches:
       - main
     types:
-      - completed
-    workflows:
-      # List all required workflow names here.
-      - 'testing'
+      - closed
 
 jobs:
-# https://github.com/marketplace/actions/tag-release-on-push-action
+  lint:
+    if: github.event.pull_request.merged == true
+    runs-on: ubuntu-latest
+    strategy:
+      max-parallel: 1
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.11
+          cache: pip
+      - name: Run pre-commit
+        run: |
+          pip install pre-commit
+          pre-commit run
+
+  test-comprehensive:
+    needs: lint
+    strategy:
+      max-parallel: 6
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
+        os:
+          - ubuntu-latest
+          - macos-latest
+          - windows-latest
+          - macos-14
+        exclude:
+          - os: macos-14
+            python-version: "3.9"
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}${{ matrix.dev }}
+      - name: Install test requirements
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e ".[testing]"
+      - name: Run tests
+        run: |
+          pytest --cov=src/pyEQL --cov-report=xml
+      - uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          file: ./coverage.xml
+
+  # https://github.com/marketplace/actions/tag-release-on-push-action
   auto-gen-release:
     name: Tag/Release on merge of labeled PRs
+    needs: test-comprehensive
     runs-on: ubuntu-latest
     env:
       GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
