### 2024-03-05 739e9ba
alpha: fix incorrect calculation
