### 2024-03-05 e262092
Docs: add warning about Apple M1 support
