+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## [0.13.0] - 2024-3-05
+## [0.13.0] - 2024-03-05
 
 ### Fixed
 
