### 2024-03-05 258bc92
CI: consolidate tagging and PyPi
