### 2024-03-06 #6295 @rkingsbury
@lucydot all items complete! The version is `v0.14.0` and the Zenodo DOI is https://doi.org/10.5281/zenodo.10783921 (for this specific version). As I'm sure you know, Zenodo also provides an "all versions" DOI.
