### 2024-03-06 #6295 @lucydot
@rkingsbury  - the title of the Zenodo release needs to **exactly** match that in the JOSS paper - can you update?
