### 2024-03-06 #6295 @lucydot
One more comment 😄 ...

Whilst reading through paper @rkingsbury I've spotted some adjustments to make:

- line 9 contains a double closing bracket
- there are two entries for reference `Marcellos, C. F. C.`. One points to a pre-print, and one to a Github repository. If you judge the repository link is required (ie, important enough that can't be followed via the pre-print) then I recommend embedding this link in the main text.
- Removing references that have no unique identifier (usually DOI) and replacing with links in main text might also be better - so that would also include `phreeqpython` and `pint`.
- on line 24 to avoid the use of consecutive brackets, could rephrase "Available software such as the open source Phree..." . Or the open source could be removed entirely as you repeat further down the paragraph.
- I would also break that sentence up: "....and contain powerful capabilities for specialists. However they are not highly....."
- Replace `;` on line 31 with `.`
- there is one too many closing brackets on line 45.
- to avoid reference-brackets within phrase-brackets (lines 19 and 45) I think you can use square brackets around the reference i.e. `@citation` instead of `[@citation]`

Once you make adjustments, please check they are as expected using `@editorialbot generate pdf`.
