### 2024-03-07 #6295 @rkingsbury
> Hello all, a quick message to say I will be out of office from tomorrow, and will check back here on the 18th.

OK, thanks for letting me know @lucydot ! I like your idea about bumping the version to 1.0 (I was thinking about this even before you posted). There are a few small changes I want to make before I do so, but I'll take care of those soon and let you know when ready.

One question - after I make the requested corrections to the manuscript, approximately how long does final acceptance usually take? I ask because I am going to feature pyEQL in a conference talk the week of the 18th, and I'd love to be able to say it has been accepted.
