+++ docs/joss/paper.bib
@@ -85,14 +85,6 @@
       archivePrefix={arXiv},
       primaryClass={physics.chem-ph}
 }
-@misc{pyequion2,
-  author = {Caio Felippe Curitiba Marcellos and Gerson Dias da Silva J{\'u}nior and Elvis do A. Soares and F{\'a}bio Ramos and Amaro Gomes Barreto},
-  title = {PyEquIon: A Python Package For Automatic Speciation Calculations of Aqueous Electrolyte Solutions},
-  year = {2021},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/pyequion/pyequion2}
-}
 @misc{phreeqpython,
   author = {Vitens, N.V.},
   title = {Phreeqpython},
