### 2024-03-07 #6295 @rkingsbury
- [x] line 9 contains a double closing bracket
- [x] there are two entries for reference `Marcellos, C. F. C.`. One points to a pre-print, and one to a Github repository. If you judge the repository link is required (ie, important enough that can't be followed via the pre-print) then I recommend embedding this link in the main text.
- [x] Removing references that have no unique identifier (usually DOI) and replacing with links in main text might also be better - so that would also include `phreeqpython` and `pint`.
- [x] on line 24 to avoid the use of consecutive brackets, could rephrase "Available software such as the open source Phree..." . Or the open source could be removed entirely as you repeat further down the paragraph.
- [x]  I would also break that sentence up: "....and contain powerful capabilities for specialists. However they are not highly....."
- [x] Replace `;` on line 31 with `.`
- [x]  there is one too many closing brackets on line 45.
- [x] to avoid reference-brackets within phrase-brackets (lines 19 and 45) I think you can use square brackets around the reference i.e. `@citation` instead of `[@citation]`
- [ ] Once you make adjustments, please check they are as expected using `@editorialbot generate pdf`.
