+++ requirements/macos-latest_py3.9_extras.txt
@@ -24,9 +24,9 @@ black==24.2.0
     # via pyEQL (setup.py)
 blinker==1.7.0
     # via flask
-boto3==1.34.46
+boto3==1.34.59
     # via maggma
-botocore==1.34.46
+botocore==1.34.59
     # via
     #   boto3
     #   s3transfer
@@ -48,9 +48,9 @@ click==8.1.7
     #   uvicorn
 contourpy==1.2.0
     # via matplotlib
-coverage[toml]==7.4.2
+coverage[toml]==7.4.3
     # via pytest-cov
-cryptography==42.0.4
+cryptography==42.0.5
     # via paramiko
 cycler==0.12.1
     # via matplotlib
@@ -69,7 +69,7 @@ exceptiongroup==1.2.0
     # via
     #   anyio
     #   pytest
-fastapi==0.109.2
+fastapi==0.110.0
     # via maggma
 filelock==3.13.1
     # via
@@ -93,11 +93,11 @@ idna==3.6
     #   requests
 imagesize==1.4.1
     # via sphinx
-importlib-metadata==7.0.1
+importlib-metadata==7.0.2
     # via
     #   flask
     #   sphinx
-importlib-resources==6.1.1
+importlib-resources==6.1.3
     # via matplotlib
 iniconfig==2.0.0
     # via pytest
@@ -120,11 +120,11 @@ jsonschema-specifications==2023.12.1
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
-latexcodec==2.0.1
+latexcodec==3.0.0
     # via pybtex
 linkify-it-py==2.0.3
     # via myst-parser
-maggma==0.63.2
+maggma==0.63.4
     # via pyEQL (setup.py)
 markdown-it-py==3.0.0
     # via
@@ -145,16 +145,16 @@ mongogrant==0.3.3
     # via maggma
 mongomock==4.1.2
     # via maggma
-monty==2024.2.2
+monty==2024.2.26
     # via
     #   maggma
     #   pyEQL (setup.py)
     #   pymatgen
 mpmath==1.3.0
     # via sympy
-msgpack==1.0.7
+msgpack==1.0.8
     # via maggma
-mypy==1.8.0
+mypy==1.9.0
     # via pyEQL (setup.py)
 mypy-extensions==1.0.0
     # via
@@ -190,7 +190,7 @@ packaging==23.2
     #   tox
 palettable==3.3.3
     # via pymatgen
-pandas==2.2.0
+pandas==2.2.1
     # via pymatgen
 paramiko==3.4.0
     # via sshtunnel
@@ -222,12 +222,12 @@ pybtex==0.24.0
     # via pymatgen
 pycparser==2.21
     # via cffi
-pydantic==2.6.1
+pydantic==2.6.3
     # via
     #   fastapi
     #   maggma
     #   pydantic-settings
-pydantic-core==2.16.2
+pydantic-core==2.16.3
     # via pydantic
 pydantic-settings==2.2.1
     # via maggma
@@ -237,25 +237,25 @@ pygments==2.17.2
     # via
     #   rich
     #   sphinx
-pymatgen==2024.2.20
+pymatgen==2024.3.1
     # via pyEQL (setup.py)
-pymongo==4.6.1
+pymongo==4.6.2
     # via
     #   maggma
     #   mongogrant
 pynacl==1.5.0
     # via paramiko
-pyparsing==3.1.1
+pyparsing==3.1.2
     # via
     #   matplotlib
     #   periodictable
-pytest==8.0.1
+pytest==8.1.1
     # via
     #   pyEQL (setup.py)
     #   pytest-cov
 pytest-cov==4.1.0
[…]
