+++ requirements/windows-latest_py3.10.txt
@@ -18,9 +18,9 @@ bcrypt==4.1.2
     # via paramiko
 blinker==1.7.0
     # via flask
-boto3==1.34.46
+boto3==1.34.59
     # via maggma
-botocore==1.34.46
+botocore==1.34.59
     # via
     #   boto3
     #   s3transfer
@@ -43,7 +43,7 @@ colorama==0.4.6
     #   tqdm
 contourpy==1.2.0
     # via matplotlib
-cryptography==42.0.4
+cryptography==42.0.5
     # via paramiko
 cycler==0.12.1
     # via matplotlib
@@ -53,7 +53,7 @@ dnspython==2.6.1
     #   pymongo
 exceptiongroup==1.2.0
     # via anyio
-fastapi==0.109.2
+fastapi==0.110.0
     # via maggma
 flask==3.0.2
     # via mongogrant
@@ -85,9 +85,9 @@ jsonschema-specifications==2023.12.1
     # via jsonschema
 kiwisolver==1.4.5
     # via matplotlib
-latexcodec==2.0.1
+latexcodec==3.0.0
     # via pybtex
-maggma==0.63.2
+maggma==0.63.4
     # via pyEQL (setup.py)
 markupsafe==2.1.5
     # via
@@ -99,14 +99,14 @@ mongogrant==0.3.3
     # via maggma
 mongomock==4.1.2
     # via maggma
-monty==2024.2.2
+monty==2024.2.26
     # via
     #   maggma
     #   pyEQL (setup.py)
     #   pymatgen
 mpmath==1.3.0
     # via sympy
-msgpack==1.0.7
+msgpack==1.0.8
     # via maggma
 networkx==3.2.1
     # via pymatgen
@@ -130,7 +130,7 @@ packaging==23.2
     #   plotly
 palettable==3.3.3
     # via pymatgen
-pandas==2.2.0
+pandas==2.2.1
     # via pymatgen
 paramiko==3.4.0
     # via sshtunnel
@@ -148,30 +148,30 @@ pybtex==0.24.0
     # via pymatgen
 pycparser==2.21
     # via cffi
-pydantic==2.6.1
+pydantic==2.6.3
     # via
     #   fastapi
     #   maggma
     #   pydantic-settings
-pydantic-core==2.16.2
+pydantic-core==2.16.3
     # via pydantic
 pydantic-settings==2.2.1
     # via maggma
 pydash==7.0.7
     # via maggma
-pymatgen==2024.2.20
+pymatgen==2024.3.1
     # via pyEQL (setup.py)
-pymongo==4.6.1
+pymongo==4.6.2
     # via
     #   maggma
     #   mongogrant
 pynacl==1.5.0
     # via paramiko
-pyparsing==3.1.1
+pyparsing==3.1.2
     # via
     #   matplotlib
     #   periodictable
-python-dateutil==2.8.2
+python-dateutil==2.9.0.post0
     # via
     #   botocore
     #   maggma
@@ -215,10 +215,9 @@ sentinels==1.0.0
     # via mongomock
 six==1.16.0
     # via
-    #   latexcodec
     #   pybtex
     #   python-dateutil
-sniffio==1.3.0
+sniffio==1.3.1
     # via anyio
 spglib==2.3.1
     # via pymatgen
@@ -236,7 +235,7 @@ tqdm==4.66.2
     # via
     #   maggma
     #   pymatgen
-typing-extensions==4.9.0
+typing-extensions==4.10.0
     # via
     #   anyio
     #   fastapi
@@ -253,7 +252,7 @@ urllib3==2.0.7
     # via
     #   botocore
     #   requests
-uvicorn==0.27.1
+uvicorn==0.28.0
     # via maggma
 werkzeug==3.0.1
     # via flask
