### 2024-03-13 16ffffb
Solution: fixes failures when oxi state is unknown
