### 2024-03-13 1b6f0ec
NativeEOS: ensure equilibrate does not change solvent mass
