### 2024-03-13 6e091c4
overhaul logging system; adjust message levels
