+++ src/pyEQL/solution.py
@@ -8,6 +8,7 @@ pyEQL Solution Class.
 
 from __future__ import annotations
 
+import logging
 import math
 import os
 import warnings
@@ -30,7 +31,6 @@ from pyEQL.activity_correction import _debye_parameter_activity, _debye_paramete
 from pyEQL.engines import EOS, IdealEOS, NativeEOS, PhreeqcEOS
 
 # logging system
-from pyEQL.logging_system import logger
 from pyEQL.salt_ion_match import Salt
 from pyEQL.solute import Solute
 from pyEQL.utils import FormulaDict, create_water_substance, standardize_formula
@@ -59,6 +59,7 @@ class Solution(MSONable):
         engine: EOS | Literal["native", "ideal", "phreeqc"] = "native",
         database: str | Path | Store | None = None,
         default_diffusion_coeff: float = 1.6106e-9,
+        log_level: Literal["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"] = "ERROR",
     ):
         """
         Instantiate a Solution from a composition.
@@ -103,6 +104,8 @@ class Solution(MSONable):
             engine: Electrolyte modeling engine to use. See documentation for details on the available engines.
             database: path to a .json file (str or Path) or maggma Store instance that
                 contains serialized SoluteDocs. `None` (default) will use the built-in pyEQL database.
+            log_level: Log messages of this or higher severity will be printed to stdout. Defaults to 'ERROR', meaning
+                that ERROR and CRITICAL messages will be shown, while WARNING, INFO, and DEBUG messages are not.
             default_diffusion_coeff: Diffusion coefficient value in m^2/s to use in
                 calculations when there is no diffusion coefficient for a species in the database. This affects several
                 important property calculations including conductivity and transport number, which are related to the
@@ -125,8 +128,20 @@ class Solution(MSONable):
             Temperature: 293.150 K
             Components: ['H2O(aq)', 'H[+1]', 'OH[-1]', 'Na[+1]', 'Cl[-1]']
         """
-        # create a logger attached to this class
-        # self.logger = logging.getLogger(type(self).__name__)
+        # create a logger and attach it to this class
+        self.log_level = log_level.upper()
+        self.logger = logging.getLogger(type(self).__name__)
+        self.logger.setLevel(self.log_level)
+        # this line prevents duplicate handlers
+        if not self.logger.handlers:
+            # use rich for pretty log formatting, if installed
+            try:
+                from rich.logging import RichHandler
+
+                sh = RichHandler(rich_tracebacks=True)
+            except ImportError:
+                sh = logging.StreamHandler()
+            self.logger.addHandler(sh)
 
         # per-instance cache of get_property and other calls that do not depend
         # on composition
@@ -174,13 +189,13 @@ class Solution(MSONable):
             db_store = IonDB
         elif isinstance(database, (str, Path)):
             db_store = JSONStore(str(database), key="formula")
-            logger.info(f"Created maggma JSONStore from .json file {database}")
+            self.logger.debug(f"Created maggma JSONStore from .json file {database}")
         else:
             db_store = database
         self.database = db_store
         """`Store` instance containing the solute property database."""
         self.database.connect()
-        logger.info(f"Connected to property database {self.database!s}")
+        self.logger.debug(f"Connected to property database {self.database!s}")
 
         # set the equation of state engine
         self._engine = engine
@@ -209,7 +224,7 @@ class Solution(MSONable):
         # TODO - do I need the ability to specify the solvent mass?
         # # raise an error if the solvent volume has also been given
         # if volume_set is True:
-        #     logger.error(
+        #     self.logger.error(
         #         "Solvent volume and mass cannot both be specified. Calculating volume based on solvent mass."
[…]
