+++ tests/test_phreeqc.py
@@ -210,7 +210,7 @@ def test_equilibrate(s1, s2, s5_pH, s6_Ca, caplog):
     # test log message if there is a species not present in the phreeqc database
     s_zr = Solution({"Zr+4": "0.05 mol/kg", "Na+": "0.05 mol/kg", "Cl-": "0.1 mol/kg"}, engine="phreeqc")
     totzr = s_zr.get_total_amount("Zr", "mol")
-    with caplog.at_level(logging.INFO, "pyEQL.logging_system"):
+    with caplog.at_level(logging.INFO, "pyEQL.engines"):
         s_zr.equilibrate()
         assert "likely absent from its database" in caplog.text
         assert "Zr[+4]" in s_zr.components

[…]
