### 2024-03-13 296be67
Solution.get_total_amount: fix error on ppm unit
