+++ src/pyEQL/solution.py
@@ -624,9 +624,7 @@ class Solution(MSONable):
         else:
             # TODO - fall back to the Jones-Dole model! There are currently no eyring parameters in the database!
             # proceed with the coefficients equal to zero and log a warning
-            self.logger.warning(
-                "Viscosity coefficients for %s not found. Viscosity will be approximate." % salt.formula
-            )
+            self.logger.warning(f"Viscosity coefficients for {salt.formula} not found. Viscosity will be approximate.")
             G_123 = G_23 = 0
 
         # get the kinematic viscosity of water, returned by IAPWS in m2/s
@@ -975,7 +973,7 @@ class Solution(MSONable):
             -1 * ureg.R * self.temperature / partial_molar_volume_water * math.log(self.get_water_activity())
         )
         self.logger.debug(
-            f"Computed osmotic pressure of solution as {osmotic_pressure} Pa at T= {self.temperature} degrees C"
+            f"Calculated osmotic pressure of solution as {osmotic_pressure} Pa at T= {self.temperature} degrees C"
         )
         return osmotic_pressure.to("Pa")
 
@@ -1122,7 +1120,7 @@ class Solution(MSONable):
                     oxi_states = self.get_property(s, "oxi_state_guesses")
                     oxi_state = oxi_states.get(el, UNKNOWN_OXI_STATE)
                 except (TypeError, IndexError):
-                    self.logger.error(f"Guessing oxi states failed for {s}. Assigning '{UNKNOWN_OXI_STATE}'")
+                    self.logger.error(f"No oxidation state found for element {el}. Assigning '{UNKNOWN_OXI_STATE}'")
                     oxi_state = UNKNOWN_OXI_STATE
                 key = f"{el}({oxi_state})"
                 if d.get(key):
@@ -1232,9 +1230,7 @@ class Solution(MSONable):
             # mw = ureg.Quantity(self.get_property(self.solvent_name, "molecular_weight"))
             mw = self.get_property(self.solvent, "molecular_weight")
             if mw is None:
-                raise ValueError(
-                    f"Molecular weight for solvent {self.solvent} not found in database. This is required to proceed."
-                )
+                raise ValueError(f"Molecular weight for solvent {self.solvent} not found in database. Cannot proceed.")
             target_mol = target_mass.to("g") / mw.to("g/mol")
             self.components[self.solvent] = target_mol.magnitude
 
@@ -1382,12 +1378,12 @@ class Solution(MSONable):
         """
         # raise an error if a negative amount is specified
         if ureg.Quantity(amount).magnitude < 0:
-            self.logger.error(f"Negative amount specified for solute {solute}. Concentration not changed.")
+            raise ValueError(f"Negative amount specified for solute {solute}. Concentration not changed.")
 
         # if units are given on a per-volume basis,
         # iteratively solve for the amount of solute that will preserve the
         # original volume and result in the desired concentration
-        elif ureg.Quantity(amount).dimensionality in (
+        if ureg.Quantity(amount).dimensionality in (
             "[substance]/[length]**3",
             "[mass]/[length]**3",
         ):
@@ -1742,7 +1738,7 @@ class Solution(MSONable):
             # get the molal-scale activity coefficient from the EOS engine
             molal = self.engine.get_activity_coefficient(solution=self, solute=solute)
         except (ValueError, ZeroDivisionError):
-            self.logger.error("Calculation unsuccessful. Returning unit activity coefficient.")
+            self.logger.error("Calculation unsuccessful. Returning unit activity coefficient.", exc_info=True)
             return ureg.Quantity(1, "dimensionless")
 
         # if necessary, convert the activity coefficient to another scale, and return the result
[…]
