### 2024-03-14 16029e7
replace maggma ref with URL
