### 2024-03-14 b0c49a6
rm 2nd maggma link
