### 2024-03-14 #6295 @rkingsbury
@lucydot final .pdf is ready for your review once you're back!

I plan to issue a `v1.0.0` in coming days, which we can set as the version of record. I will post here when that release is up.
