### 2024-03-16 KingsburyLab/pyEQL#109 @xiaoxiaozhu123
It's because phreeqpython is not compatible with arm64 architecture.
Use a python installer or a miniconda installer for x86 architecture. They works on my M2 Mac.
DO NOT use the arm64 versions for macs with M1/M2 chips. Phreeqpython just won't work.
