### 2024-03-17 KingsburyLab/pyEQL#109 @xiaoxiaozhu123
> > The problem had been confusing me for weeks.... It's because phreeqpython is not compatible with arm64 architecture. SOLUTION IS SIMPLE: If you want to run Solutions with the phreeqc engine, make a python env using a [miniconda Intel-x86 installer](https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.pkg) (despite that it's designated for intel-chip macs, it's compatible with Apple Silicon. The only side effect is loss of performance, but python is already too slow for you to notice this...). At least pyEQL works on my M2 Mac now. DO NOT use the arm64 version installer for macs with M1/M2 chips. Phreeqpython just won't work.
>
> Thanks for posting this solution @xiaoxiaozhu123 ! I'm actually going to be issuing a 1.0 release in the next 1-2 days and I'd love to update the documentation with this. Would you be able to suggest some edits / additions to the "Warning" box [on this page](https://pyeql.readthedocs.io/en/latest/installation.html#pip-install)? Feel free to post here or open a PR if you're comfortable doing that.

As far as I know,  It's impossible toggle the current architecture with just a python script. The only thing we can do is to inform users about compatibility issues and recommend actions they can take to adjust their environment.
So it'll be good to execute a platform inspection in pyEQL's __init__.py. Here is a GPT demonstration:
```
import platform
import warnings

machine = platform.machine()  # Get the machine type (e.g., 'x86_64', 'arm64')

if machine == 'arm64':
    platform_warning_msg = ...
    warnings.warn(platform_warning_msg)
    disable_functions_related_to_phreeqc() # I guess here'll be similar statements in the future release...
```
