### 2024-03-17 90885ae
use png rather than svg logo in readme
