### 2024-03-17 KingsburyLab/pyEQL#109 @rkingsbury
Yes, I do something similar to this in the unit tests right now. There is currently an error message that is raised, which I think is adequate as a technical solution for the time being.

```
            logger.error(
                "OSError encountered when trying to instantiate phreeqpython. Most likely this means you"
                " are running on an architecture that is not supported by PHREEQC, such as Apple M1/M2 chips."
                " pyEQL will work, but equilibrate() will have no effect."
            )
```

I have just updated the documentation page, adding the following so that other users can try your workaround. This appears in the Warning box that mentions the challenges with M1/M2.

```
NOTE: Some users have reported being able to use `phreeqpython` functions by installing
an x86 version of `conda`/`miniconda`. See the issue report for more details.
```
