### 2024-03-17 0f81d20
logging updates
