### 2024-03-17 bf98d45
move logging setup to init; add tests
