+++ src/pyEQL/__init__.py
@@ -5,6 +5,8 @@ and performing chemical thermodynamics computations.
 :copyright: 2013-2024 by Ryan S. Kingsbury
 :license: LGPL, see LICENSE for more details.
 """
+
+import logging
 from importlib.metadata import PackageNotFoundError, version  # pragma: no cover
 from importlib.resources import files
 
@@ -19,6 +21,12 @@ except PackageNotFoundError:  # pragma: no cover
     __version__ = "unknown"
 finally:
     del version, PackageNotFoundError
+
+# logging
+logger = logging.getLogger("pyEQL")
+logger.setLevel(logging.WARNING)
+logger.addHandler(logging.NullHandler())
+
 
 # Units handling
 # per the pint documentation, it's important that pint and its associated Unit
