### 2024-03-17 #6295 @rkingsbury
(Updating the figure with a slightly modernized logo, in honor of v1.0)
