### 2024-03-17 26e09ca
smaller logo in docs
