+++ src/pyEQL/functions.py
@@ -5,10 +5,8 @@ pyEQL functions that take Solution objects as inputs or return Solution objects.
 :license: LGPL, see LICENSE for more details.
 
 """
-import math
-from typing import Literal
 
-from monty.dev import deprecated
+import math
 
 from pyEQL import Solution, ureg
 from pyEQL.logging_system import logger
@@ -264,157 +262,3 @@ def donnan_eql(solution: Solution, fixed_charge: str):
 
     # return the equilibrated solution
     return donnan_soln
-
-
-@deprecated(
-    message="mix() is deprecated and will be removed in the next release! You can now mix solutions using the addition operator, e.g. s_mix = s1 + s2."
-)
-def mix(s1, s2):  # pragma: no cover
-    """
-    Mix two solutions together.
-
-    Args:
-        s1, s2: The two solutions to be mixed.
-
-    Returns:
-        A Solution object that represents the result of mixing s1 and s2.
-
-    Notes:
-        The initial volume of the mixed solution is set as the sum of the volumes of s1 and s2. The pressure and
-        temperature are volume-weighted averages. The pH and pE values are currently APPROXIMATE because they are
-        calculated assuming H+ and e- mix conservatively (i.e., the mixing process does not incorporate any
-        equilibration reactions or buffering). Such support is planned in a future release.
-    """
-    return s1 + s2
-
-
-@deprecated(
-    message="autogenerate() is deprecated and will be removed in the next release! Use Solution.from_preset() instead.)"
-)
-def autogenerate(
-    solution: Literal["seawater", "rainwater", "wastewater", "urine", "Ringers lactate", "normal saline"]
-):  # pragma: no cover
-    """
-    This method provides a quick way to create Solution objects representing
-    commonly-encountered solutions, such as seawater, rainwater, and wastewater.
-
-    Parameters
-    ----------
-    solution : str
-                String representing the desired solution
-                Valid entries are 'seawater', 'rainwater',
-                'wastewater',and 'urine'
-
-    Returns
-    -------
-    Solution
-        A pyEQL Solution object.
-
-    Notes
-    -----
-    The following sections explain the different solution options:
-
-    - '' - empty solution, equivalent to pyEQL.Solution()
-    - 'rainwater' - pure water in equilibrium with atmospheric CO2 at pH 6
-    - 'seawater' or 'SW'- Standard Seawater. See Table 4 of the Reference for Composition [1]_
-    - 'wastewater' or 'WW' - medium strength domestic wastewater. See Table 3-18 of [2]_
-    - 'urine' - typical human urine. See Table 3-15 of [2]_
-    - 'normal saline' or 'NS' - normal saline solution used in medicine [3]_
-    - 'Ringers lacatate' or 'RL' - Ringer's lactate solution used in medicine [4]_
-
-    References:
-    ----------
-        .. [1] Millero, Frank J. "The composition of Standard Seawater and the definition of
-            the Reference-Composition Salinity Scale." *Deep-sea Research. Part I* 55(1), 2008, 50-72.
-
-        .. [2] Metcalf & Eddy, Inc. et al. *Wastewater Engineering: Treatment and Resource Recovery*, 5th Ed.
-                McGraw-Hill, 2013.
-
-        .. [3] https://en.wikipedia.org/wiki/Saline_(medicine)
-
[…]
