### 2024-03-17 aea688f
changelog, add DeprecationWarning
