+++ CHANGELOG.md
@@ -11,13 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Solution.__init__`: new keyword argument `log_level` allows user to control the verbosity of log messages by setting
   the level (e.g., ERROR, WARNING, etc.) that will be shown in stdout.
+- Docs: added a note about a workaround for Apple M1/M2 Macs proposed by @xiaoxiaozhu123
 
 ### Changed
 
 - `Solution.__init__`: The deprecated format for specifying solutes (e.g., `[["Na+", "0.5 mol/L]]`)
   which previously only generated log warning message, now raises a `DeprecationWarning`. Use dict-style input (e.g.,
   `{"Na+":"0.5 mol/L"}`) instead.
-- Docs: added a note about a workaround for Apple M1/M2 Macs proposed by @xiaoxiaozhu123
 - New logo! Updated the `pyEQL` logo (for the first time in 9 years!) to address an obsolete font in the .svg
   and modernize the design.
 
