### 2024-03-17 #6295 @rkingsbury
@lucydot v1.0.0 is out and the .pdf is finalized. Please set that as the version of record, and let me know if anything else is needed. Thank you again to you and the reviewers for your feedback!
