### 2024-03-18 #6295 @lucydot
@rkingsbury one small change - can you add a doi field to the arxiv entry? It should then create a url at doi.org rather than pointing to the arxiv.org.
