### 2024-03-18 b8e30a6
rm extra .bib; add doi to arxiv paper
