+++ docs/joss/paper.bib
@@ -83,31 +83,8 @@
       year={2021},
       eprint={2101.07246},
       archivePrefix={arXiv},
-      primaryClass={physics.chem-ph}
-}
-@misc{phreeqpython,
-  author = {Vitens, N.V.},
-  title = {Phreeqpython},
-  year = {2016},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/Vitens/phreeqpython}
-}
-@misc{pint,
-  author = {Hernan Grecco},
-  title = {Pint: makes units easy},
-  year = {2012},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/hgrecco/pint}
-}
-@misc{maggma,
-  author = {Materials Project, The},
-  title = {Maggma: A files-to-API data pipeline for scientific applications},
-  year = {2017},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/materialsproject/maggma}
+      primaryClass={physics.chem-ph},
+      doi={10.48550/arXiv.2101.07246}
 }
 @misc{gwb,
   author = {Aqueous Solutions, LLC},
