### 2024-03-19 #6295 @lucydot
Excellent - I'll recommend acceptance now. Thanks for submitting your work to JOSS, and hope you have a great conference!
