+++ docs/joss/paper.bib
@@ -75,7 +75,7 @@
   pages = {1653--1663},
   author = {Scott R. Charlton and David L. Parkhurst},
   title = {Modules based on the geochemical model {PHREEQC} for use in scripting and programming languages},
-  journal = {Computers {\&}amp$\mathsemicolon$ Geosciences}
+  journal = {Computers \& Geosciences}
 }
 @misc{marcellos2021pyequion,
       title={PyEquIon: A Python Package For Automatic Speciation Calculations of Aqueous Electrolyte Solutions},
