### 2024-03-19 #6295 @rkingsbury
> @rkingsbury - the problem is with the `charlton2011` entry in the bib file. I think it should be easy enough to fix by using `\&` for the ampersand.

Fixed.
