### 2024-03-21 #6295 @lucydot
@rkingsbury - I'm not quite sure why your paper doesn't appear under Active. It may be because we have recommended accept (it might remove it off active paper database at that point - though I'm not sure of mechanisms under the hood).

It seems that one of the ISBN's is too short; 9 rather than length 10. Looking at .bib file, I think the issue is with Parkhurst2013 entry; could you double check the ISBN/entry?
