### 2024-03-21 746fa72
update PHREEQC manual bib
