+++ docs/joss/paper.bib
@@ -1,13 +1,11 @@
-@incollection{Parkhurst2013,
-  title = {Description of {{Input}} and {{Examples}} for {{PHREEQC Version}} 3 — {{A Computer Program}} for {{Speciation}} , {{Batch-Reaction}} , {{One-Dimensional Transport}} , and {{Inverse Geochemical Calculations Chapter}} 43 Of},
-  booktitle = {Modeling {{Techniques}}, {{Book}} 6},
-  author = {Parkhurst, David L. and Appelo, C.a.J.},
-  date = {2013},
-  eprint = {11908328},
-  eprinttype = {pmid},
-  pages = {1--678},
-  abstract = {Description of Input and Examples for PHREEQC Version 3—A Computer Program for Speciation, Batch-Reaction, One-Dimensional Transport, and Inverse Geochemical Calculations},
-  isbn = {0092-332X}
+@techreport{Parkhurst2013,
+  title = {Description of input and examples for PHREEQC version 3: a computer program for speciation,  batch-reaction,  one-dimensional transport,  and inverse geochemical calculations},
+  ISSN = {2328-7055},
+  url = {http://dx.doi.org/10.3133/tm6A43},
+  DOI = {10.3133/tm6a43},
+  institution = {US Geological Survey},
+  author = {Parkhurst,  David L. and Appelo,  C.A.J.},
+  year = {2013}
 }
 @article{rowland_ProgressAqueousSolution_2019,
   title = {Progress in {{Aqueous Solution Modelling}}: {{Better Data}} and {{Better Interfaces}}},
