### 2024-03-21 #6295 @rkingsbury
> It seems that one of the ISBN's is too short; 9 rather than length 10. Looking at .bib file, I think the issue is with Parkhurst2013 entry; could you double check the ISBN/entry?

OK, I updated this .bib entry to use a different style that includes a DOI instead of an ISBN. Why do you think this didn't show up in the previous try of `recommend-accept`? It would be nice to see all these errors at once instead of solving them one at a time.
