### 2024-03-22 #6295 @kthyng
Paper:
* type first paragraph 3rd page: "to maintain tranpra..."
* Architecture > The Solution class > "constituitive" is that a word?
