### 2024-03-23 #6295 @rkingsbury
>     * type first paragraph 3rd page: "to maintain tranpra..."

Fixed

>     * Architecture > The Solution class > "constituitive" is that a word?

Yes, this is a word. See https://en.wikipedia.org/wiki/Constitutive_equation
