### 2024-03-25 94f2434
Fix a missing close parenthesis in docstring
